% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_activity}
\alias{add_activity}
\title{Add activity}
\usage{
add_activity(data_tb, neither_1L_chr = "ERROR")
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{neither_1L_chr}{Neither (a character vector of length one), Default: 'ERROR'}
}
\value{
Data (a tibble)
}
\description{
add_activity() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add activity. The function returns Data (a tibble).
}
