% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_age_to_project_dss}
\alias{add_age_to_project_dss}
\title{Add age to project datasets}
\usage{
add_age_to_project_dss(
  project_dss_ls,
  age_1L_chr = "Age",
  drop_1L_lgl = FALSE,
  date_of_birth_1L_chr = "date_of_birth",
  index_date_1L_chr = "onboarding_date",
  what_chr = c("contacts", "outcomes", "overview")
)
}
\arguments{
\item{project_dss_ls}{Project datasets (a list)}

\item{age_1L_chr}{Age (a character vector of length one), Default: 'Age'}

\item{drop_1L_lgl}{Drop (a logical vector of length one), Default: FALSE}

\item{date_of_birth_1L_chr}{Date of birth (a character vector of length one), Default: 'date_of_birth'}

\item{index_date_1L_chr}{Index date (a character vector of length one), Default: 'onboarding_date'}

\item{what_chr}{What (a character vector), Default: c("contacts", "outcomes", "overview")}
}
\value{
Project datasets (a list)
}
\description{
add_age_to_project_dss() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add age to project datasets. The function returns Project datasets (a list).
}
\keyword{internal}
