% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_clients_to_summary}
\alias{add_clients_to_summary}
\title{Add clients to summary}
\usage{
add_clients_to_summary(
  summaries_ls,
  arrange_1L_chr = character(0),
  reference_1L_chr = "Status quo",
  weeks_1L_int = 26
)
}
\arguments{
\item{summaries_ls}{Summaries (a list)}

\item{arrange_1L_chr}{Arrange (a character vector of length one), Default: character(0)}

\item{reference_1L_chr}{Reference (a character vector of length one), Default: 'Status quo'}

\item{weeks_1L_int}{Weeks (an integer vector of length one), Default: 26}
}
\value{
Summaries (a list)
}
\description{
add_clients_to_summary() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add clients to summary. The function returns Summaries (a list).
}
\keyword{internal}
