% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cost_calculations}
\alias{add_cost_calculations}
\title{Add cost calculations}
\usage{
add_cost_calculations(
  data_tb,
  inputs_ls,
  add_fixed_1L_lgl = FALSE,
  add_logic_fn = identity,
  add_offsets_1L_lgl = FALSE,
  add_variable_1L_lgl = TRUE,
  base_for_rates_int = 1L,
  offsets_chr = character(0),
  variable_unit_1L_chr = "Minutes"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{inputs_ls}{Inputs (a list)}

\item{add_fixed_1L_lgl}{Add fixed (a logical vector of length one), Default: FALSE}

\item{add_logic_fn}{Add logic (a function), Default: identity}

\item{add_offsets_1L_lgl}{Add offsets (a logical vector of length one), Default: FALSE}

\item{add_variable_1L_lgl}{Add variable (a logical vector of length one), Default: TRUE}

\item{base_for_rates_int}{Base for rates (an integer vector), Default: 1}

\item{offsets_chr}{Offsets (a character vector), Default: character(0)}

\item{variable_unit_1L_chr}{Variable unit (a character vector of length one), Default: 'Minutes'}
}
\value{
Data (a tibble)
}
\description{
add_cost_calculations() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add cost calculations. The function returns Data (a tibble).
}
