% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cost_effectiveness}
\alias{add_cost_effectiveness}
\title{Add cost effectiveness}
\usage{
add_cost_effectiveness(
  data_tb,
  cost_1L_chr = "Cost",
  dominance_1L_chr = character(0),
  effect_1L_chr = "QALYs",
  icer_1L_chr = character(0),
  suffix_1L_chr = "",
  threshold_1L_dbl = 96000
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{cost_1L_chr}{Cost (a character vector of length one), Default: 'Cost'}

\item{dominance_1L_chr}{Dominance (a character vector of length one), Default: character(0)}

\item{effect_1L_chr}{Effect (a character vector of length one), Default: 'QALYs'}

\item{icer_1L_chr}{Icer (a character vector of length one), Default: character(0)}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{threshold_1L_dbl}{Threshold (a double vector of length one), Default: 96000}
}
\value{
Data (a tibble)
}
\description{
add_cost_effectiveness() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add cost effectiveness. The function returns Data (a tibble).
}
\keyword{internal}
