% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cost_effectiveness_stats}
\alias{add_cost_effectiveness_stats}
\title{Add cost effectiveness statistics}
\usage{
add_cost_effectiveness_stats(data_tb, threshold_1L_dbl = 96000)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{threshold_1L_dbl}{Threshold (a double vector of length one), Default: 96000}
}
\value{
Data (a tibble)
}
\description{
add_cost_effectiveness_stats() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add cost effectiveness statistics. The function returns Data (a tibble).
}
\keyword{internal}
