% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cost_offsets}
\alias{add_cost_offsets}
\title{Add cost offsets}
\usage{
add_cost_offsets(
  data_tb,
  inputs_ls,
  offsets_chr,
  add_logic_fn = identity,
  base_for_rates_int = 1L
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{inputs_ls}{Inputs (a list)}

\item{offsets_chr}{Offsets (a character vector)}

\item{add_logic_fn}{Add logic (a function), Default: identity}

\item{base_for_rates_int}{Base for rates (an integer vector), Default: 1}
}
\value{
Data (a tibble)
}
\description{
add_cost_offsets() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add cost offsets. The function returns Data (a tibble).
}
