% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_costs_to_summary}
\alias{add_costs_to_summary}
\title{Add costs to summary}
\usage{
add_costs_to_summary(
  summaries_ls,
  processed_ls,
  periods_1L_int = 26,
  arrange_1L_chr = character(0),
  reference_1L_chr = "Status quo",
  tfmn_fn = identity,
  type_1L_chr = c("change_scenario", "change_sq", "scenario", "sq")
)
}
\arguments{
\item{summaries_ls}{Summaries (a list)}

\item{processed_ls}{Processed (a list)}

\item{periods_1L_int}{Periods (an integer vector of length one), Default: 26}

\item{arrange_1L_chr}{Arrange (a character vector of length one), Default: character(0)}

\item{reference_1L_chr}{Reference (a character vector of length one), Default: 'Status quo'}

\item{tfmn_fn}{Transformation (a function), Default: identity}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("change_scenario", "change_sq", "scenario", "sq")}
}
\value{
Summaries (a list)
}
\description{
add_costs_to_summary() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add costs to summary. The function returns Summaries (a list).
}
\keyword{internal}
