% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_draws_from_pool}
\alias{add_draws_from_pool}
\title{Add draws from pool}
\usage{
add_draws_from_pool(
  draws_tb,
  inputs_ls,
  iterations_1L_int,
  scale_1L_int = 10L,
  seed_1L_int = 1000L
)
}
\arguments{
\item{draws_tb}{Draws (a tibble)}

\item{inputs_ls}{Inputs (a list)}

\item{iterations_1L_int}{Iterations (an integer vector of length one)}

\item{scale_1L_int}{Scale (an integer vector of length one), Default: 10}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}
}
\value{
Draws (a tibble)
}
\description{
add_draws_from_pool() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add draws from pool. The function returns Draws (a tibble).
}
\keyword{internal}
