% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_enter_model_event}
\alias{add_enter_model_event}
\title{Add enter model event}
\usage{
add_enter_model_event(
  X_Ready4useDyad,
  arm_1L_chr,
  draws_tb,
  horizon_dtm = lubridate::years(1),
  iterations_int = 1:100L,
  modifiable_chr = character(0),
  start_dtm = Sys.Date(),
  tfmn_ls = NULL,
  tx_duration_dtm = lubridate::weeks(12)
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{arm_1L_chr}{Arm (a character vector of length one)}

\item{draws_tb}{Draws (a tibble)}

\item{horizon_dtm}{Horizon (a date vector), Default: lubridate::years(1)}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100L}

\item{modifiable_chr}{Modifiable (a character vector), Default: character(0)}

\item{start_dtm}{Start (a date vector), Default: Sys.Date()}

\item{tfmn_ls}{Transformation (a list), Default: NULL}

\item{tx_duration_dtm}{Treatment duration (a date vector), Default: lubridate::weeks(12)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_enter_model_event() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add enter model event. The function is called for its side effects and does not return a value.
}
\keyword{internal}
