% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_episode_start}
\alias{add_episode_start}
\title{Add episode start}
\usage{
add_episode_start(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_episode_start() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add episode start. The function is called for its side effects and does not return a value.
}
