% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_eq5d_from_draws}
\alias{add_eq5d_from_draws}
\title{Add EQ5D from draws}
\usage{
add_eq5d_from_draws(
  X_Ready4useDyad,
  correspondences_r3 = ready4show::ready4show_correspondences(),
  prefix_1L_chr = "ParamEQ5DBeta",
  value_with_fn = add_eq5d_from_k10,
  var_1L_chr = "EQ5D"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{correspondences_r3}{Correspondences (a ready4 submodule), Default: ready4show::ready4show_correspondences()}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'ParamEQ5DBeta'}

\item{value_with_fn}{Value with (a function), Default: add_eq5d_from_k10}

\item{var_1L_chr}{Variable (a character vector of length one), Default: 'EQ5D'}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_eq5d_from_draws() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add eq5d from draws. The function is called for its side effects and does not return a value.
}
