% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_eq5d_from_k10}
\alias{add_eq5d_from_k10}
\title{Add EQ5D from K10}
\usage{
add_eq5d_from_k10(
  data_xx,
  correspondences_r3 = ready4show::ready4show_correspondences(),
  beta_age_1L_dbl = -0.01382,
  beta_constant_1L_dbl = 3.5222,
  beta_k10_1L_dbl = -0.06476,
  germany_1L_lgl = FALSE,
  prefix_1L_chr = "ParamEQ5DBeta",
  var_1L_chr = "EQ5D",
  source_1L_chr = c("10.1192/bjo.2018.21", "10.1192/bjp.bp.113.136036"),
  tidy_cols_1L_lgl = FALSE,
  type_1L_chr = c("internal", "external")
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{correspondences_r3}{Correspondences (a ready4 submodule), Default: ready4show::ready4show_correspondences()}

\item{beta_age_1L_dbl}{Beta age (a double vector of length one), Default: -0.01382}

\item{beta_constant_1L_dbl}{Beta constant (a double vector of length one), Default: 3.5222}

\item{beta_k10_1L_dbl}{Beta K10 (a double vector of length one), Default: -0.06476}

\item{germany_1L_lgl}{Germany (a logical vector of length one), Default: FALSE}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'ParamEQ5DBeta'}

\item{var_1L_chr}{Variable (a character vector of length one), Default: 'EQ5D'}

\item{source_1L_chr}{Source (a character vector of length one), Default: c("10.1192/bjo.2018.21", "10.1192/bjp.bp.113.136036")}

\item{tidy_cols_1L_lgl}{Tidy columns (a logical vector of length one), Default: FALSE}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("internal", "external")}
}
\value{
Data (an output object of multiple potential types)
}
\description{
add_eq5d_from_k10() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add eq5d from k10. The function returns Data (an output object of multiple potential types).
}
