% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_iar_params}
\alias{add_iar_params}
\title{Add Initial Assessment andeferral parameters}
\usage{
add_iar_params(
  params_tb,
  comparator_int,
  model_data_ls,
  processed_ls,
  raw_mds_data_ls,
  test_1L_chr,
  comparator_1L_chr = "Comparator",
  comparator_filter_fn = identity,
  cost_1L_dbl = 0,
  intervention_1L_chr = "Intervention",
  intervention_filter_fn = identity
)
}
\arguments{
\item{params_tb}{Parameters (a tibble)}

\item{comparator_int}{Comparator (an integer vector)}

\item{model_data_ls}{Model data (a list)}

\item{processed_ls}{Processed (a list)}

\item{raw_mds_data_ls}{Raw Minimum Dataset data (a list)}

\item{test_1L_chr}{Test (a character vector of length one)}

\item{comparator_1L_chr}{Comparator (a character vector of length one), Default: 'Comparator'}

\item{comparator_filter_fn}{Comparator filter (a function), Default: identity}

\item{cost_1L_dbl}{Cost (a double vector of length one), Default: 0}

\item{intervention_1L_chr}{Intervention (a character vector of length one), Default: 'Intervention'}

\item{intervention_filter_fn}{Intervention filter (a function), Default: identity}
}
\value{
Parameters (a tibble)
}
\description{
add_iar_params() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add initial assessment andeferral parameters. The function returns Parameters (a tibble).
}
