% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_icer}
\alias{add_icer}
\title{Add icer}
\usage{
add_icer(
  data_tb,
  cost_1L_chr = "Cost",
  effect_1L_chr = "QALYs",
  suffix_1L_chr = ""
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{cost_1L_chr}{Cost (a character vector of length one), Default: 'Cost'}

\item{effect_1L_chr}{Effect (a character vector of length one), Default: 'QALYs'}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}
}
\value{
Data (a tibble)
}
\description{
add_icer() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add icer. The function returns Data (a tibble).
}
