% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_imputed_data}
\alias{add_imputed_data}
\title{Add imputed data}
\usage{
add_imputed_data(
  X_Ready4useDyad,
  Y_Ready4useDyad = ready4use::Ready4useDyad(),
  characteristics_chr = c("platform", "clinic_type", "gender", "employment_status",
    "clinic_state", "clinic_postcode", "role_type", "treatment_stage"),
  extras_chr = character(0),
  impute_age_1L_lgl = FALSE,
  method_1L_chr = "rf",
  treatment_status_1L_int = 0L,
  x_is_z_1L_lgl = FALSE
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{characteristics_chr}{Characteristics (a character vector), Default: c("platform", "clinic_type", "gender", "employment_status", "clinic_state",
"clinic_postcode", "role_type", "treatment_stage")}

\item{extras_chr}{Extras (a character vector), Default: character(0)}

\item{impute_age_1L_lgl}{Impute age (a logical vector of length one), Default: FALSE}

\item{method_1L_chr}{Method (a character vector of length one), Default: 'rf'}

\item{treatment_status_1L_int}{Treatment status (an integer vector of length one), Default: 0}

\item{x_is_z_1L_lgl}{X is z (a logical vector of length one), Default: FALSE}
}
\value{
Z (A dataset and data dictionary pair.)
}
\description{
add_imputed_data() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add imputed data. The function is called for its side effects and does not return a value.
}
\keyword{internal}
