% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_ineligible}
\alias{add_ineligible}
\title{Add ineligible}
\usage{
add_ineligible(
  X_Ready4useDyad,
  ineligible_1L_chr = character(0),
  post_fn = identity,
  pre_fn = identity,
  type_1L_chr = c("filter", "reset")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{ineligible_1L_chr}{Ineligible (a character vector of length one), Default: character(0)}

\item{post_fn}{Post (a function), Default: identity}

\item{pre_fn}{Pre (a function), Default: identity}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("filter", "reset")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_ineligible() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add ineligible. The function is called for its side effects and does not return a value.
}
\keyword{internal}
