% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_iteration_values_set}
\alias{add_iteration_values_set}
\title{Add iteration values set}
\usage{
add_iteration_values_set(
  X_Ready4useDyad,
  value_with_fn,
  value_with_args_ls = NULL,
  tidy_cols_1L_lgl = TRUE
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{value_with_fn}{Value with (a function)}

\item{value_with_args_ls}{Value with arguments (a list), Default: NULL}

\item{tidy_cols_1L_lgl}{Tidy columns (a logical vector of length one), Default: TRUE}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_iteration_values_set() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add iteration values set. The function is called for its side effects and does not return a value.
}
\keyword{internal}
