% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_k10_event}
\alias{add_k10_event}
\title{Add K10 event}
\usage{
add_k10_event(
  X_Ready4useDyad,
  adjustment_1L_dbl = 0,
  k10_mdl = NULL,
  iterations_int = 1:100L,
  params_tb = make_project_params_tb(),
  sensitivities_ls = make_sensitivities_ls(),
  suffix_1L_chr = character(0),
  tfmn_ls = make_class_tfmns(T),
  type_1L_chr = c("Model", "Project", "Table")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{adjustment_1L_dbl}{Adjustment (a double vector of length one), Default: 0}

\item{k10_mdl}{K10 (a model), Default: NULL}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100L}

\item{params_tb}{Parameters (a tibble), Default: make_project_params_tb()}

\item{sensitivities_ls}{Sensitivities (a list), Default: make_sensitivities_ls()}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: character(0)}

\item{tfmn_ls}{Transformation (a list), Default: make_class_tfmns(T)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("Model", "Project", "Table")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_k10_event() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add k10 event. The function is called for its side effects and does not return a value.
}
\keyword{internal}
