% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_leave_model_event}
\alias{add_leave_model_event}
\title{Add leave model event}
\usage{
add_leave_model_event(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_leave_model_event() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add leave model event. The function is called for its side effects and does not return a value.
}
