% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_mds_minutes_totals}
\alias{add_mds_minutes_totals}
\title{Add Minimum Dataset minutes totals}
\usage{
add_mds_minutes_totals(
  services_tb,
  add_chr = c("Contacts", "Use"),
  type_1L_chr = c("both", "total", "prop")
)
}
\arguments{
\item{services_tb}{Services (a tibble)}

\item{add_chr}{Add (a character vector), Default: c("Contacts", "Use")}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("both", "total", "prop")}
}
\value{
Services (a tibble)
}
\description{
add_mds_minutes_totals() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add minimum dataset minutes totals. The function returns Services (a tibble).
}
\keyword{internal}
