% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_mds_org_vars}
\alias{add_mds_org_vars}
\title{Add Minimum Dataset organisation variables}
\usage{
add_mds_org_vars(
  data_tb,
  provider_lup_tb,
  phn_code_1L_chr = "PHN_code",
  phn_name_1L_chr = "PHN_area_name"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{provider_lup_tb}{Provider lookup table (a tibble)}

\item{phn_code_1L_chr}{Primary Health Network code (a character vector of length one), Default: 'PHN_code'}

\item{phn_name_1L_chr}{Primary Health Network name (a character vector of length one), Default: 'PHN_area_name'}
}
\value{
Data (a tibble)
}
\description{
add_mds_org_vars() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add minimum dataset organisation variables. The function returns Data (a tibble).
}
