% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_mds_program_vars}
\alias{add_mds_program_vars}
\title{Add Minimum Dataset program variables}
\usage{
add_mds_program_vars(
  data_tb,
  processed_ls,
  program_services_lup,
  program_true_chr,
  program_type_ls,
  provider_lup_tb,
  add_start_date_1L_lgl = TRUE,
  filter_fn = identity,
  mature_after_dtm = lubridate::years(1)
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{processed_ls}{Processed (a list)}

\item{program_services_lup}{Program services (a lookup table)}

\item{program_true_chr}{Program true (a character vector)}

\item{program_type_ls}{Program type (a list)}

\item{provider_lup_tb}{Provider lookup table (a tibble)}

\item{add_start_date_1L_lgl}{Add start date (a logical vector of length one), Default: TRUE}

\item{filter_fn}{Filter (a function), Default: identity}

\item{mature_after_dtm}{Mature after (a date vector), Default: lubridate::years(1)}
}
\value{
Data (a tibble)
}
\description{
add_mds_program_vars() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add minimum dataset program variables. The function returns Data (a tibble).
}
\keyword{internal}
