% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_minutes_event}
\alias{add_minutes_event}
\title{Add minutes event}
\usage{
add_minutes_event(
  X_Ready4useDyad,
  add_dependency_1L_lgl = T,
  minutes_mdl = NULL,
  iterations_int = 1:100L,
  fraction_1L_dbl = numeric(0),
  var_1L_chr = "Minutes"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{add_dependency_1L_lgl}{Add dependency (a logical vector of length one), Default: T}

\item{minutes_mdl}{Minutes (a model), Default: NULL}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100L}

\item{fraction_1L_dbl}{Fraction (a double vector of length one), Default: numeric(0)}

\item{var_1L_chr}{Variable (a character vector of length one), Default: 'Minutes'}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_minutes_event() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add minutes event. The function is called for its side effects and does not return a value.
}
\keyword{internal}
