% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_non_iar}
\alias{add_non_iar}
\title{Add non Initial Assessment andeferral}
\usage{
add_non_iar(X_Ready4useDyad, arms_for_iar_adjustment_chr = character(0))
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{arms_for_iar_adjustment_chr}{Arms for Initial Assessment andeferral adjustment (a character vector), Default: character(0)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_non_iar() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add non initial assessment andeferral. The function is called for its side effects and does not return a value.
}
