% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_outcome_change_schedule}
\alias{add_outcome_change_schedule}
\title{Add outcome change schedule}
\usage{
add_outcome_change_schedule(X_Ready4useDyad, step_dtm = lubridate::weeks(0))
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{step_dtm}{Step (a date vector), Default: lubridate::weeks(0)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_outcome_change_schedule() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add outcome change schedule. The function is called for its side effects and does not return a value.
}
