% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_outcome_sensitivity}
\alias{add_outcome_sensitivity}
\title{Add outcome sensitivity}
\usage{
add_outcome_sensitivity(
  X_Ready4useDyad,
  outcome_1L_chr,
  sensitivities_ls = make_sensitivities_ls(),
  tfmn_fn = NULL,
  tfmn_ls = make_class_tfmns(T)
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{outcome_1L_chr}{Outcome (a character vector of length one)}

\item{sensitivities_ls}{Sensitivities (a list), Default: make_sensitivities_ls()}

\item{tfmn_fn}{Transformation (a function), Default: NULL}

\item{tfmn_ls}{Transformation (a list), Default: make_class_tfmns(T)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_outcome_sensitivity() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add outcome sensitivity. The function is called for its side effects and does not return a value.
}
\keyword{internal}
