% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_outcome_time_vars}
\alias{add_outcome_time_vars}
\title{Add outcome time variables}
\usage{
add_outcome_time_vars(
  Y_Ready4useDyad,
  outcome_1L_chr,
  add_adjustments_1L_lgl = FALSE,
  fup_var_1L_chr = character(0),
  follow_up_1L_int = integer(0),
  maintain_for_1L_int = 0L
)
}
\arguments{
\item{Y_Ready4useDyad}{PARAM_DESCRIPTION}

\item{outcome_1L_chr}{Outcome (a character vector of length one)}

\item{add_adjustments_1L_lgl}{Add adjustments (a logical vector of length one), Default: FALSE}

\item{fup_var_1L_chr}{Follow-up variable (a character vector of length one), Default: character(0)}

\item{follow_up_1L_int}{Follow up (an integer vector of length one), Default: integer(0)}

\item{maintain_for_1L_int}{Maintain for (an integer vector of length one), Default: 0}
}
\value{
Y (A dataset and data dictionary pair.)
}
\description{
add_outcome_time_vars() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add outcome time variables. The function is called for its side effects and does not return a value.
}
