% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_outcomes_update}
\alias{add_outcomes_update}
\title{Add outcomes update}
\usage{
add_outcomes_update(
  X_Ready4useDyad,
  assert_1L_lgl,
  k10_mdl,
  k10_var_1L_chr,
  iterations_int,
  params_tb,
  sensitivities_ls,
  tfmn_ls,
  tx_prefix_1L_chr,
  update_1L_int,
  utilities_chr,
  utility_fns_ls,
  types_chr = c("Model", "Function")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{assert_1L_lgl}{Assert (a logical vector of length one)}

\item{k10_mdl}{K10 (a model)}

\item{k10_var_1L_chr}{K10 variable (a character vector of length one)}

\item{iterations_int}{Iterations (an integer vector)}

\item{params_tb}{Parameters (a tibble)}

\item{sensitivities_ls}{Sensitivities (a list)}

\item{tfmn_ls}{Transformation (a list)}

\item{tx_prefix_1L_chr}{Treatment prefix (a character vector of length one)}

\item{update_1L_int}{Update (an integer vector of length one)}

\item{utilities_chr}{Utilities (a character vector)}

\item{utility_fns_ls}{Utility functions (a list)}

\item{types_chr}{Types (a character vector), Default: c("Model", "Function")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_outcomes_update() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add outcomes update. The function is called for its side effects and does not return a value.
}
\keyword{internal}
