% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_project_2_cost_sa_2}
\alias{add_project_2_cost_sa_2}
\title{Add project 2 cost sensitivity analysis 2}
\usage{
add_project_2_cost_sa_2(
  X_Ready4useDyad,
  arms_for_intervention_costs_chr = "Intervention",
  disciplines_chr = make_disciplines(),
  suffix_1L_chr = "_S2",
  ...
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{arms_for_intervention_costs_chr}{Arms for intervention costs (a character vector), Default: 'Intervention'}

\item{disciplines_chr}{Disciplines (a character vector), Default: make_disciplines()}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: '_S2'}

\item{...}{Additional arguments}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_project_2_cost_sa_2() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add project 2 cost sensitivity analysis 2. The function is called for its side effects and does not return a value.
}
