% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_project_2_customisation}
\alias{add_project_2_customisation}
\title{Add project 2 customisation}
\usage{
add_project_2_customisation(
  X_Ready4useDyad,
  arms_for_non_helpseeking_chr = character(0),
  arms_for_iar_adjustment_chr = character(0),
  reset_date_1L_lgl = TRUE
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{arms_for_non_helpseeking_chr}{Arms for non helpseeking (a character vector), Default: character(0)}

\item{arms_for_iar_adjustment_chr}{Arms for Initial Assessment andeferral adjustment (a character vector), Default: character(0)}

\item{reset_date_1L_lgl}{Reset date (a logical vector of length one), Default: TRUE}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_project_2_customisation() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add project 2 customisation. The function is called for its side effects and does not return a value.
}
\keyword{internal}
