% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_project_2_model_data}
\alias{add_project_2_model_data}
\title{Add project 2 model data}
\usage{
add_project_2_model_data(
  model_data_ls,
  sample_ls,
  intervention_1L_chr = "Intervention",
  cut_off_date_1L_chr = "2025-01-01"
)
}
\arguments{
\item{model_data_ls}{Model data (a list)}

\item{sample_ls}{Sample (a list)}

\item{intervention_1L_chr}{Intervention (a character vector of length one), Default: 'Intervention'}

\item{cut_off_date_1L_chr}{Cut off date (a character vector of length one), Default: '2025-01-01'}
}
\value{
Model data (a list)
}
\description{
add_project_2_model_data() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add project 2 model data. The function returns Model data (a list).
}
\keyword{internal}
