% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_project_2_model_wrap_up}
\alias{add_project_2_model_wrap_up}
\title{Add project 2 model wrap up}
\usage{
add_project_2_model_wrap_up(
  X_Ready4useDyad,
  arms_for_intervention_costs_chr,
  arms_for_offsets_chr = character(0),
  defaults_ls = NULL,
  disciplines_chr,
  inputs_ls,
  iterations_int,
  outcome_1L_chr = "K10",
  sensitivities_ls,
  tfmn_ls,
  tx_prefix_1L_chr,
  update_1L_int = 2,
  utilities_chr,
  utility_fns_ls
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{arms_for_intervention_costs_chr}{Arms for intervention costs (a character vector)}

\item{arms_for_offsets_chr}{Arms for offsets (a character vector), Default: character(0)}

\item{defaults_ls}{Defaults (a list), Default: NULL}

\item{disciplines_chr}{Disciplines (a character vector)}

\item{inputs_ls}{Inputs (a list)}

\item{iterations_int}{Iterations (an integer vector)}

\item{outcome_1L_chr}{Outcome (a character vector of length one), Default: 'K10'}

\item{sensitivities_ls}{Sensitivities (a list)}

\item{tfmn_ls}{Transformation (a list)}

\item{tx_prefix_1L_chr}{Treatment prefix (a character vector of length one)}

\item{update_1L_int}{Update (an integer vector of length one), Default: 2}

\item{utilities_chr}{Utilities (a character vector)}

\item{utility_fns_ls}{Utility functions (a list)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_project_2_model_wrap_up() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add project 2 model wrap up. The function is called for its side effects and does not return a value.
}
\keyword{internal}
