% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_project_2_offsets}
\alias{add_project_2_offsets}
\title{Add project 2 offsets}
\usage{
add_project_2_offsets(X_Ready4useDyad, arms_for_offsets_chr = character(0))
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{arms_for_offsets_chr}{Arms for offsets (a character vector), Default: character(0)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_project_2_offsets() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add project 2 offsets. The function is called for its side effects and does not return a value.
}
\keyword{internal}
