% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_project_offset_logic}
\alias{add_project_offset_logic}
\title{Add project offset logic}
\usage{
add_project_offset_logic(data_tb)
}
\arguments{
\item{data_tb}{Data (a tibble)}
}
\value{
Data (a tibble)
}
\description{
add_project_offset_logic() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add project offset logic. The function returns Data (a tibble).
}
\keyword{internal}
