% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_project_treatment_change}
\alias{add_project_treatment_change}
\title{Add project treatment change}
\usage{
add_project_treatment_change(
  ds_tb,
  arrange_by_id_lgl = T,
  change_var_nm_1L_chr = "treatment_change",
  suffix_1L_chr = "_previous",
  timepoint_1L_chr = "MeasurementWeek",
  timepoint_bl_1L_chr = "Week0",
  uid_1L_chr = "UID",
  var_nm_1L_chr = "treatment_status",
  wide_1L_lgl = F
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{arrange_by_id_lgl}{Arrange by identity (a logical vector), Default: T}

\item{change_var_nm_1L_chr}{Change variable name (a character vector of length one), Default: 'treatment_change'}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: '_previous'}

\item{timepoint_1L_chr}{Timepoint (a character vector of length one), Default: 'MeasurementWeek'}

\item{timepoint_bl_1L_chr}{Timepoint baseline (a character vector of length one), Default: 'Week0'}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'UID'}

\item{var_nm_1L_chr}{Variable name (a character vector of length one), Default: 'treatment_status'}

\item{wide_1L_lgl}{Wide (a logical vector of length one), Default: F}
}
\value{
Dataset (a tibble)
}
\description{
add_project_treatment_change() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add project treatment change. The function returns Dataset (a tibble).
}
