% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_projected_growth}
\alias{add_projected_growth}
\title{Add projected growth}
\usage{
add_projected_growth(
  X_Ready4useDyad,
  outcome_1L_chr,
  suffix_1L_chr,
  proportion_1L_dbl = 0.2,
  tfmn_fn = identity,
  ...
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{outcome_1L_chr}{Outcome (a character vector of length one)}

\item{suffix_1L_chr}{Suffix (a character vector of length one)}

\item{proportion_1L_dbl}{Proportion (a double vector of length one), Default: 0.2}

\item{tfmn_fn}{Transformation (a function), Default: identity}

\item{...}{Additional arguments}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_projected_growth() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add projected growth. The function is called for its side effects and does not return a value.
}
\keyword{internal}
