% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_projected_maintenance}
\alias{add_projected_maintenance}
\title{Add projected maintenance}
\usage{
add_projected_maintenance(
  X_Ready4useDyad,
  outcome_1L_chr,
  suffix_1L_chr,
  tfmn_fn = identity,
  ...
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{outcome_1L_chr}{Outcome (a character vector of length one)}

\item{suffix_1L_chr}{Suffix (a character vector of length one)}

\item{tfmn_fn}{Transformation (a function), Default: identity}

\item{...}{Additional arguments}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_projected_maintenance() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add projected maintenance. The function is called for its side effects and does not return a value.
}
