% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_regressions}
\alias{add_regressions}
\title{Add regressions}
\usage{
add_regressions(
  regressions_ls,
  model_1L_int = integer(0),
  fn_args_ls = list(),
  model_fn = NULL,
  named_1L_lgl = FALSE,
  X_Ready4useDyad = ready4use::Ready4useDyad(),
  type_1L_chr = c("candidates", "tests", "models"),
  what_1L_chr = c("AQoL6D", "CHU9D", "K10", "Minutes", "Treatments", "Tx_Waitlist",
    "Tx_Treatment", "Tx_Discharged")
)
}
\arguments{
\item{regressions_ls}{Regressions (a list)}

\item{model_1L_int}{Model (an integer vector of length one), Default: integer(0)}

\item{fn_args_ls}{Function arguments (a list), Default: list()}

\item{model_fn}{Model (a function), Default: NULL}

\item{named_1L_lgl}{Named (a logical vector of length one), Default: FALSE}

\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("candidates", "tests", "models")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("AQoL6D", "CHU9D", "K10", "Minutes", "Treatments", "Tx_Waitlist",
"Tx_Treatment", "Tx_Discharged")}
}
\value{
Regressions (a list)
}
\description{
add_regressions() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add regressions. The function returns Regressions (a list).
}
\keyword{internal}
