% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_severity_cuts}
\alias{add_severity_cuts}
\title{Add severity cuts}
\usage{
add_severity_cuts(
  X_Ready4useDyad,
  add_severity_1L_lgl = TRUE,
  severity_fn = make_k10_severity_cuts,
  severity_var_1L_chr = "k10_start"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{add_severity_1L_lgl}{Add severity (a logical vector of length one), Default: TRUE}

\item{severity_fn}{Severity (a function), Default: make_k10_severity_cuts}

\item{severity_var_1L_chr}{Severity variable (a character vector of length one), Default: 'k10_start'}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_severity_cuts() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add severity cuts. The function is called for its side effects and does not return a value.
}
\keyword{internal}
