% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_simulated_data}
\alias{add_simulated_data}
\title{Add simulated data}
\usage{
add_simulated_data(
  model_mdl,
  var_1L_chr,
  Y_Ready4useDyad,
  iterations_int = 1:100L,
  join_with_chr = character(0),
  tfmn_1L_chr = "NTF",
  type_1L_chr = c("first", "second", "third", "fourth"),
  what_1L_chr = c("old", "new")
)
}
\arguments{
\item{model_mdl}{Model (a model)}

\item{var_1L_chr}{Variable (a character vector of length one)}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100L}

\item{join_with_chr}{Join with (a character vector), Default: character(0)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("first", "second", "third", "fourth")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("old", "new")}
}
\value{
Y (A dataset and data dictionary pair.)
}
\description{
add_simulated_data() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add simulated data. The function is called for its side effects and does not return a value.
}
\keyword{internal}
