% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_time_to_event}
\alias{add_time_to_event}
\title{Add time to event}
\usage{
add_time_to_event(
  X_Ready4useDyad,
  event_1L_chr,
  schedule_args_ls = list(),
  schedule_fn = NULL,
  step_dtm = lubridate::days(0)
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{event_1L_chr}{Event (a character vector of length one)}

\item{schedule_args_ls}{Schedule arguments (a list), Default: list()}

\item{schedule_fn}{Schedule (a function), Default: NULL}

\item{step_dtm}{Step (a date vector), Default: lubridate::days(0)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_time_to_event() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add time to event. The function is called for its side effects and does not return a value.
}
