% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_treatment_status}
\alias{add_treatment_status}
\title{Add treatment status}
\usage{
add_treatment_status(
  data_xx,
  arrange_by_1L_chr = c("category", "name"),
  ctg_1L_chr = "Service",
  group_by_1L_chr = character(0),
  source_vars_chr = c("treatment_stage", "stage_of_treatment"),
  three_levels_1L_lgl = FALSE,
  type_1L_int = 1L:2L,
  update_dict_1L_lgl = TRUE,
  var_1L_chr = "treatment_status"
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{arrange_by_1L_chr}{Arrange by (a character vector of length one), Default: c("category", "name")}

\item{ctg_1L_chr}{Category (a character vector of length one), Default: 'Service'}

\item{group_by_1L_chr}{Group by (a character vector of length one), Default: character(0)}

\item{source_vars_chr}{Source variables (a character vector), Default: c("treatment_stage", "stage_of_treatment")}

\item{three_levels_1L_lgl}{Three levels (a logical vector of length one), Default: FALSE}

\item{type_1L_int}{Type (an integer vector of length one), Default: 1L:2L}

\item{update_dict_1L_lgl}{Update dictionary (a logical vector of length one), Default: TRUE}

\item{var_1L_chr}{Variable (a character vector of length one), Default: 'treatment_status'}
}
\value{
Data (an output object of multiple potential types)
}
\description{
add_treatment_status() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add treatment status. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
