% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_unset_vars}
\alias{add_unset_vars}
\title{Add unset variables}
\usage{
add_unset_vars(data_tb, var_names_chr, value_xx = 0)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{var_names_chr}{Variable names (a character vector)}

\item{value_xx}{Value (an output object of multiple potential types), Default: 0}
}
\value{
Data (a tibble)
}
\description{
add_unset_vars() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add unset variables. The function returns Data (a tibble).
}
\keyword{internal}
