% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_utility_event}
\alias{add_utility_event}
\title{Add utility event}
\usage{
add_utility_event(
  X_Ready4useDyad,
  add_qalys_1L_lgl = FALSE,
  add_sensitivity_1L_lgl = FALSE,
  adjustment_1L_dbl = 0,
  follow_up_1L_int = integer(0),
  iterations_int = 1:100L,
  maintain_for_1L_int = 0L,
  models_ls = NULL,
  sensitivities_ls = make_sensitivities_ls(),
  simulate_1L_lgl = TRUE,
  tfmn_ls = NULL,
  tidy_1L_lgl = TRUE,
  utilities_chr = c("AQoL6D", "CHU9D"),
  type_1L_chr = c("Model", "Project"),
  what_1L_chr = c("old", "new")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{add_qalys_1L_lgl}{Add Quality Adjusted Life Years (a logical vector of length one), Default: FALSE}

\item{add_sensitivity_1L_lgl}{Add sensitivity (a logical vector of length one), Default: FALSE}

\item{adjustment_1L_dbl}{Adjustment (a double vector of length one), Default: 0}

\item{follow_up_1L_int}{Follow up (an integer vector of length one), Default: integer(0)}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100L}

\item{maintain_for_1L_int}{Maintain for (an integer vector of length one), Default: 0}

\item{models_ls}{Models (a list), Default: NULL}

\item{sensitivities_ls}{Sensitivities (a list), Default: make_sensitivities_ls()}

\item{simulate_1L_lgl}{Simulate (a logical vector of length one), Default: TRUE}

\item{tfmn_ls}{Transformation (a list), Default: NULL}

\item{tidy_1L_lgl}{Tidy (a logical vector of length one), Default: TRUE}

\item{utilities_chr}{Utilities (a character vector), Default: c("AQoL6D", "CHU9D")}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("Model", "Project")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("old", "new")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_utility_event() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add utility event. The function is called for its side effects and does not return a value.
}
\keyword{internal}
