% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_wrap_up_date}
\alias{add_wrap_up_date}
\title{Add wrap up date}
\usage{
add_wrap_up_date(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_wrap_up_date() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add wrap up date. The function is called for its side effects and does not return a value.
}
\keyword{internal}
