% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author-MimicRepos}
\alias{author-MimicRepos}
\alias{author,MimicRepos-method}
\alias{author-MimicConfiguration}
\alias{author,MimicConfiguration-method}
\title{Author and save files}
\usage{
\S4method{author}{MimicRepos}(
  x,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N"),
  suffix_1L_chr = "",
  what_1L_chr = c("sim_ws_dirs_chr"),
  ...
)

\S4method{author}{MimicConfiguration}(
  x,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N"),
  unlink_1L_lgl = FALSE,
  what_1L_chr = c("draws"),
  Y_MimicRepos = MimicRepos()
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{what_1L_chr}{What (a character vector of length one), Default: c("draws")}

\item{...}{Additional arguments}

\item{unlink_1L_lgl}{Unlink (a logical vector of length one), Default: FALSE}

\item{Y_MimicRepos}{PARAM_DESCRIPTION, Default: MimicRepos()}
}
\value{
No return value, called for side effects.

No return value, called for side effects.
}
\description{
author method applied to MimicRepos

author method applied to MimicConfiguration
}
