% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_aqol8d_from_k10}
\alias{calculate_aqol8d_from_k10}
\title{Calculate Assessment of Quality of Life Eight Dimension from K10}
\usage{
calculate_aqol8d_from_k10(
  age_1L_dbl,
  k10_1L_dbl,
  norway_1L_lgl = FALSE,
  source_1L_chr = c("10.1192/bjp.bp.113.136036")
)
}
\arguments{
\item{age_1L_dbl}{Age (a double vector of length one)}

\item{k10_1L_dbl}{K10 (a double vector of length one)}

\item{norway_1L_lgl}{Norway (a logical vector of length one), Default: FALSE}

\item{source_1L_chr}{Source (a character vector of length one), Default: c("10.1192/bjp.bp.113.136036")}
}
\value{
an Assessment of Quality of Life Eight Dimension (a double vector of length one)
}
\description{
calculate_aqol8d_from_k10() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate assessment of quality of life eight dimension from k10. The function returns an Assessment of Quality of Life Eight Dimension (a double vector of length one).
}
