% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_eq5d_from_k10}
\alias{calculate_eq5d_from_k10}
\title{Calculate EQ5D from K10}
\usage{
calculate_eq5d_from_k10(
  age_1L_dbl,
  k10_1L_dbl,
  germany_1L_lgl = FALSE,
  beta_age_1L_dbl = -0.01382,
  beta_constant_1L_dbl = 3.5222,
  beta_k10_1L_dbl = -0.06476,
  source_1L_chr = c("10.1192/bjo.2018.21", "10.1192/bjp.bp.113.136036")
)
}
\arguments{
\item{age_1L_dbl}{Age (a double vector of length one)}

\item{k10_1L_dbl}{K10 (a double vector of length one)}

\item{germany_1L_lgl}{Germany (a logical vector of length one), Default: FALSE}

\item{beta_age_1L_dbl}{Beta age (a double vector of length one), Default: -0.01382}

\item{beta_constant_1L_dbl}{Beta constant (a double vector of length one), Default: 3.5222}

\item{beta_k10_1L_dbl}{Beta K10 (a double vector of length one), Default: -0.06476}

\item{source_1L_chr}{Source (a character vector of length one), Default: c("10.1192/bjo.2018.21", "10.1192/bjp.bp.113.136036")}
}
\value{
EQ5D (a double vector of length one)
}
\description{
calculate_eq5d_from_k10() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate eq5d from k10. The function returns EQ5D (a double vector of length one).
}
\keyword{internal}
