% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_erp_for_areas}
\alias{calculate_erp_for_areas}
\title{Calculate Estimatedesident Population for areas}
\usage{
calculate_erp_for_areas(
  raw_erp_tb = NULL,
  path_1L_chr = character(0),
  areas_chr,
  count_1L_chr = character(0),
  name_1L_chr = character(0),
  summarise_1L_lgl = TRUE,
  var_1L_chr = "LGA"
)
}
\arguments{
\item{raw_erp_tb}{Raw Estimatedesident Population (a tibble), Default: NULL}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{areas_chr}{Areas (a character vector)}

\item{count_1L_chr}{Count (a character vector of length one), Default: character(0)}

\item{name_1L_chr}{Name (a character vector of length one), Default: character(0)}

\item{summarise_1L_lgl}{Summarise (a logical vector of length one), Default: TRUE}

\item{var_1L_chr}{Variable (a character vector of length one), Default: 'LGA'}
}
\value{
Estimatedesident Population for areas (a double vector)
}
\description{
calculate_erp_for_areas() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate estimatedesident population for areas. The function returns Estimatedesident Population for areas (a double vector).
}
\keyword{internal}
