% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_offset_prob_proxy}
\alias{calculate_offset_prob_proxy}
\title{Calculate offset probability proxy}
\usage{
calculate_offset_prob_proxy(
  area_erp_1L_dbl,
  exposed_1L_dbl,
  rate_1L_dbl,
  risk_1L_dbl,
  denominator_1L_dbl = 1e+05,
  time_1L_dbl = 1
)
}
\arguments{
\item{area_erp_1L_dbl}{Area Estimatedesident Population (a double vector of length one)}

\item{exposed_1L_dbl}{Exposed (a double vector of length one)}

\item{rate_1L_dbl}{Rate (a double vector of length one)}

\item{risk_1L_dbl}{Risk (a double vector of length one)}

\item{denominator_1L_dbl}{Denominator (a double vector of length one), Default: 1e+05}

\item{time_1L_dbl}{Time (a double vector of length one), Default: 1}
}
\value{
Probability proxy (a double vector of length one)
}
\description{
calculate_offset_prob_proxy() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate offset probability proxy. The function returns Probability proxy (a double vector of length one).
}
