% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_duplicated_measures}
\alias{get_duplicated_measures}
\title{Get duplicated measures}
\usage{
get_duplicated_measures(
  data_tb,
  group_by_chr,
  type_1L_chr = c("table", "uid"),
  uid_1L_chr = "episode_key",
  ungroup_1L_lgl = FALSE
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{group_by_chr}{Group by (a character vector)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("table", "uid")}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'episode_key'}

\item{ungroup_1L_lgl}{Ungroup (a logical vector of length one), Default: FALSE}
}
\value{
Duplicates (an output object of multiple potential types)
}
\description{
get_duplicated_measures() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get duplicated measures. The function returns Duplicates (an output object of multiple potential types).
}
