% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_pooled}
\alias{get_pooled}
\title{Get pooled}
\usage{
get_pooled(
  pooled_ls,
  what_1L_chr,
  as_1L_chr = c("vector", "histogram", "summary"),
  n_1L_int = 5000,
  seed_1L_int = 2001L,
  resample_1L_lgl = TRUE
)
}
\arguments{
\item{pooled_ls}{Pooled (a list)}

\item{what_1L_chr}{What (a character vector of length one)}

\item{as_1L_chr}{As (a character vector of length one), Default: c("vector", "histogram", "summary")}

\item{n_1L_int}{N (an integer vector of length one), Default: 5000}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 2001}

\item{resample_1L_lgl}{Resample (a logical vector of length one), Default: TRUE}
}
\value{
Predictions (an output object of multiple potential types)
}
\description{
get_pooled() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get pooled. The function returns Predictions (an output object of multiple potential types).
}
\keyword{internal}
