% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_private_keys_ls}
\alias{get_private_keys_ls}
\title{Get private keys list}
\usage{
get_private_keys_ls(
  path_to_keys_1L_chr,
  divider_1L_chr = "//",
  what_1L_chr = c("data", "names")
)
}
\arguments{
\item{path_to_keys_1L_chr}{Path to keys (a character vector of length one)}

\item{divider_1L_chr}{Divider (a character vector of length one), Default: '//'}

\item{what_1L_chr}{What (a character vector of length one), Default: c("data", "names")}
}
\value{
Private keys (an output object of multiple potential types)
}
\description{
get_private_keys_ls() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get private keys list. The function returns Private keys (an output object of multiple potential types).
}
