% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_project_model_data}
\alias{get_project_model_data}
\title{Get project model data}
\usage{
get_project_model_data(
  model_data_ls,
  what_1L_chr,
  type_1L_chr = c("imputed", "unimputed")
)
}
\arguments{
\item{model_data_ls}{Model data (a list)}

\item{what_1L_chr}{What (a character vector of length one)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("imputed", "unimputed")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
get_project_model_data() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get project model data. The function is called for its side effects and does not return a value.
}
