% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_raw_params_data}
\alias{get_raw_params_data}
\title{Get raw parameters data}
\usage{
get_raw_params_data(
  path_to_param_data_1L_chr,
  program_fl_nm_1L_chr,
  unit_cost_fl_nm_1L_chr,
  cost_types_chr = character(0),
  divider_1L_chr = "\\\\",
  erp_fl_nm_1L_chr = character(0),
  mbs_fl_nm_1L_chr = character(0),
  mbs_sheet_1L_chr = "Table EXP.14",
  mbs_skip_1L_int = 4,
  no_care_fl_nm_1L_chr = "HILDA_K10.RDS",
  program_sheet_1L_chr,
  program_skip_1L_int = 0,
  r_dir_1L_chr = "R"
)
}
\arguments{
\item{path_to_param_data_1L_chr}{Path to parameter data (a character vector of length one)}

\item{program_fl_nm_1L_chr}{Program file name (a character vector of length one)}

\item{unit_cost_fl_nm_1L_chr}{Unit cost file name (a character vector of length one)}

\item{cost_types_chr}{Cost types (a character vector), Default: character(0)}

\item{divider_1L_chr}{Divider (a character vector of length one), Default: '\'}

\item{erp_fl_nm_1L_chr}{Estimatedesident Population file name (a character vector of length one), Default: character(0)}

\item{mbs_fl_nm_1L_chr}{Medicare Benefits Schedule file name (a character vector of length one), Default: character(0)}

\item{mbs_sheet_1L_chr}{Medicare Benefits Schedule sheet (a character vector of length one), Default: 'Table EXP.14'}

\item{mbs_skip_1L_int}{Medicare Benefits Schedule skip (an integer vector of length one), Default: 4}

\item{no_care_fl_nm_1L_chr}{No care file name (a character vector of length one), Default: 'HILDA_K10.RDS'}

\item{program_sheet_1L_chr}{Program sheet (a character vector of length one)}

\item{program_skip_1L_int}{Program skip (an integer vector of length one), Default: 0}

\item{r_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}
}
\value{
Raw parameters data (a list)
}
\description{
get_raw_params_data() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get raw parameters data. The function returns Raw parameters data (a list).
}
\keyword{internal}
