% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_regression}
\alias{get_regression}
\title{Get regression}
\usage{
get_regression(
  regressions_ls,
  what_1L_chr,
  constrained_1L_lgl = logical(0),
  model_1L_int = integer(0),
  named_1L_lgl = FALSE,
  part_1L_int = integer(0),
  report_1L_chr = c("all", "check", "compare", "confusion", "density", "estimates",
    "histogram", "scatter", "test"),
  type_1L_chr = c("candidates", "assessments", "models", "tests")
)
}
\arguments{
\item{regressions_ls}{Regressions (a list)}

\item{what_1L_chr}{What (a character vector of length one)}

\item{constrained_1L_lgl}{Constrained (a logical vector of length one), Default: logical(0)}

\item{model_1L_int}{Model (an integer vector of length one), Default: integer(0)}

\item{named_1L_lgl}{Named (a logical vector of length one), Default: FALSE}

\item{part_1L_int}{Part (an integer vector of length one), Default: integer(0)}

\item{report_1L_chr}{Report (a character vector of length one), Default: c("all", "check", "compare", "confusion", "density", "estimates",
"histogram", "scatter", "test")}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("candidates", "assessments", "models", "tests")}
}
\value{
Model (an output object of multiple potential types)
}
\description{
get_regression() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get regression. The function returns Model (an output object of multiple potential types).
}
\keyword{internal}
