% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_timestamp}
\alias{get_timestamp}
\title{Get timestamp}
\usage{
get_timestamp(
  sensitivities_ls = make_sensitivities_ls(),
  prefix_1L_chr = "_",
  what_1L_chr = c("outcomes", "costs")
)
}
\arguments{
\item{sensitivities_ls}{Sensitivities (a list), Default: make_sensitivities_ls()}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: '_'}

\item{what_1L_chr}{What (a character vector of length one), Default: c("outcomes", "costs")}
}
\value{
Timestamp (a character vector of length one)
}
\description{
get_timestamp() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get timestamp. The function returns Timestamp (a character vector of length one).
}
\keyword{internal}
