% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_import.R
\name{import_population_k10}
\alias{import_population_k10}
\title{Import population K10}
\usage{
import_population_k10(
  dir_1L_chr,
  fl_nm_1L_chr = "HILDA k10.xlsx",
  areas_chr = c("Intervention", "Matched"),
  divider_1L_chr = "\\\\"
)
}
\arguments{
\item{dir_1L_chr}{Directory (a character vector of length one)}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'HILDA k10.xlsx'}

\item{areas_chr}{Areas (a character vector), Default: c("Intervention", "Matched")}

\item{divider_1L_chr}{Divider (a character vector of length one), Default: '\'}
}
\value{
Population K10 (a tibble)
}
\description{
import_population_k10() is an Import function that reads a data object in its native format and converts it to an R object. Specifically, this function implements an algorithm to import population k10. The function returns Population K10 (a tibble).
}
\keyword{internal}
