% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_import.R
\name{import_project_data}
\alias{import_project_data}
\title{Import project data}
\usage{
import_project_data(
  path_to_private_1L_chr,
  dir_1L_chr,
  r_dir_1L_chr = "R",
  divider_1L_chr = "\\\\",
  names_ls = NULL,
  type_1L_chr = c("raw", "experts", "processed", "modelling", "pooled", "population",
    "regressions", "results", "simulation")
)
}
\arguments{
\item{path_to_private_1L_chr}{Path to private (a character vector of length one)}

\item{dir_1L_chr}{Directory (a character vector of length one)}

\item{r_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}

\item{divider_1L_chr}{Divider (a character vector of length one), Default: '\'}

\item{names_ls}{Names (a list), Default: NULL}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("raw", "experts", "processed", "modelling", "pooled", "population",
"regressions", "results", "simulation")}
}
\value{
Data (a list)
}
\description{
import_project_data() is an Import function that reads a data object in its native format and converts it to an R object. Specifically, this function implements an algorithm to import project data. The function returns Data (a list).
}
\keyword{internal}
