% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_import.R
\name{import_results_batches}
\alias{import_results_batches}
\title{Import results batches}
\usage{
import_results_batches(
  batches_1L_int = integer(0),
  dir_1L_chr = character(0),
  drop_params_1L_lgl = FALSE,
  ratify_1L_lgl = FALSE,
  suffix_1L_chr = "",
  use_chr = character(0),
  Y_MimicRepos = MimicRepos()
)
}
\arguments{
\item{batches_1L_int}{Batches (an integer vector of length one), Default: integer(0)}

\item{dir_1L_chr}{Directory (a character vector of length one)}

\item{drop_params_1L_lgl}{Drop parameters (a logical vector of length one), Default: FALSE}

\item{ratify_1L_lgl}{Ratify (a logical vector of length one), Default: FALSE}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ""}

\item{use_chr}{Use (a character vector), Default: character(0)}

\item{Y_MimicRepos}{(an instance of the MimicRepos class), Default: MimicRepos()}
}
\value{
Results (a list)
}
\description{
import_results_batches() is an Import function that reads a data object in its native format and converts it to an R object. Specifically, this function implements an algorithm to import results batches. The function returns Results (a list).
}
\keyword{internal}
