% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ingest.R
\name{ingest-MimicRepos}
\alias{ingest-MimicRepos}
\alias{ingest,MimicRepos-method}
\title{Ingest data}
\usage{
\S4method{ingest}{MimicRepos}(
  x,
  gh_token_1L_chr = "",
  key_1L_chr = NULL,
  remote_fls_chr = NA_character_,
  type_1L_chr = c("MimicInputs", "list", "element"),
  what_chr = character(0),
  ...
)
}
\arguments{
\item{x}{An object of class MimicRepos}

\item{gh_token_1L_chr}{Github token (a character vector of length one), Default: ''}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{remote_fls_chr}{Remote files (a character vector), Default: 'NA'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("MimicInputs", "list", "element")}

\item{what_chr}{What (a character vector), Default: character(0)}

\item{...}{Additional arguments}
}
\value{
Ingest (an output object of multiple potential types)
}
\description{
ingest method applied to MimicRepos
}
