% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_annual_overview}
\alias{make_annual_overview}
\title{Make annual overview}
\usage{
make_annual_overview(processed_ls)
}
\arguments{
\item{processed_ls}{Processed (a list)}
}
\value{
Annual (a tibble)
}
\description{
make_annual_overview() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make annual overview. The function returns Annual (a tibble).
}
