% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_arms_tb}
\alias{make_arms_tb}
\title{Make arms tibble}
\usage{
make_arms_tb(arms_chr = character(0), settings_ls = NULL)
}
\arguments{
\item{arms_chr}{Arms (a character vector), Default: character(0)}

\item{settings_ls}{Settings (a list), Default: NULL}
}
\value{
Arms (a tibble)
}
\description{
make_arms_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make arms tibble. The function returns Arms (a tibble).
}
\keyword{internal}
