% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aus_price_years}
\alias{make_aus_price_years}
\title{Make Australia price years}
\usage{
make_aus_price_years(start_1L_chr = "2021-22", end_1L_chr = "2024-25")
}
\arguments{
\item{start_1L_chr}{Start (a character vector of length one), Default: '2021-22'}

\item{end_1L_chr}{End (a character vector of length one), Default: '2024-25'}
}
\value{
Price indices (a double vector)
}
\description{
make_aus_price_years() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make australia price years. The function returns Price indices (a double vector).
}
