% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_batches}
\alias{make_batches}
\title{Make batches}
\usage{
make_batches(n_1L_int = 50, of_1L_int = 20)
}
\arguments{
\item{n_1L_int}{N (an integer vector of length one), Default: 50}

\item{of_1L_int}{Of (an integer vector of length one), Default: 20}
}
\value{
Batches (a list)
}
\description{
make_batches() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make batches. The function returns Batches (a list).
}
