% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_class_tfmns}
\alias{make_class_tfmns}
\title{Make class transformations}
\usage{
make_class_tfmns(force_1L_lgl = FALSE)
}
\arguments{
\item{force_1L_lgl}{Force (a logical vector of length one), Default: FALSE}
}
\value{
Transformations (a list)
}
\description{
make_class_tfmns() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make class transformations. The function returns Transformations (a list).
}
\keyword{internal}
