% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_clinic_summary}
\alias{make_clinic_summary}
\title{Make clinic summary}
\usage{
make_clinic_summary(processed_ls, type_1L_chr = serious::make_temporal_vars())
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{type_1L_chr}{Type (a character vector of length one), Default: serious::make_temporal_vars()}
}
\value{
Clinics (a tibble)
}
\description{
make_clinic_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make clinic summary. The function returns Clinics (a tibble).
}
