% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_conditional_vars}
\alias{make_conditional_vars}
\title{Make conditional variables}
\usage{
make_conditional_vars(
  outcome_1L_chr,
  follow_up_1L_int = integer(0),
  fup_var_1L_chr = character(0),
  type_1L_chr = c("end", "fup", "start", "years")
)
}
\arguments{
\item{outcome_1L_chr}{Outcome (a character vector of length one)}

\item{follow_up_1L_int}{Follow up (an integer vector of length one), Default: integer(0)}

\item{fup_var_1L_chr}{Follow-up variable (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("end", "fup", "start", "years")}
}
\value{
Variable (a character vector of length one)
}
\description{
make_conditional_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make conditional variables. The function returns Variable (a character vector of length one).
}
\keyword{internal}
