% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_configuration}
\alias{make_configuration}
\title{Make configuration}
\usage{
make_configuration(
  arms_chr,
  drop_missing_1L_lgl,
  drop_suffix_1L_chr,
  extra_draws_fn,
  horizon_dtm,
  iterations_ls,
  modifiable_chr,
  seed_1L_int,
  sensitivities_ls,
  start_dtm,
  synthesis_fn,
  transformations_ls,
  utilities_chr,
  arms_extras_ls = list()
)
}
\arguments{
\item{arms_chr}{Arms (a character vector)}

\item{drop_missing_1L_lgl}{Drop missing (a logical vector of length one)}

\item{drop_suffix_1L_chr}{Drop suffix (a character vector of length one)}

\item{extra_draws_fn}{Extra draws (a function)}

\item{horizon_dtm}{Horizon (a date vector)}

\item{iterations_ls}{Iterations (a list)}

\item{modifiable_chr}{Modifiable (a character vector)}

\item{seed_1L_int}{Seed (an integer vector of length one)}

\item{sensitivities_ls}{Sensitivities (a list)}

\item{start_dtm}{Start (a date vector)}

\item{synthesis_fn}{Synthesis (a function)}

\item{transformations_ls}{Transformations (a list)}

\item{utilities_chr}{Utilities (a character vector)}

\item{arms_extras_ls}{Arms extras (a list), Default: list()}
}
\value{
X (Configuration details for a simulation run.)
}
\description{
make_configuration() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make configuration. The function is called for its side effects and does not return a value.
}
\keyword{internal}
