% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_contacters_series}
\alias{make_contacters_series}
\title{Make contacters series}
\usage{
make_contacters_series(model_data_ls)
}
\arguments{
\item{model_data_ls}{Model data (a list)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
make_contacters_series() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make contacters series. The function is called for its side effects and does not return a value.
}
\keyword{internal}
