% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_contacters_summary}
\alias{make_contacters_summary}
\title{Make contacters summary}
\usage{
make_contacters_summary(
  processed_ls,
  as_tsibble_1L_lgl = FALSE,
  type_1L_chr = serious::make_temporal_vars()
)
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{as_tsibble_1L_lgl}{As tsibble (a logical vector of length one), Default: FALSE}

\item{type_1L_chr}{Type (a character vector of length one), Default: serious::make_temporal_vars()}
}
\value{
Contacters (a tibble)
}
\description{
make_contacters_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make contacters summary. The function returns Contacters (a tibble).
}
\keyword{internal}
