% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_disciplines}
\alias{make_disciplines}
\title{Make disciplines}
\usage{
make_disciplines(
  arrange_1L_chr = c("default", "semi", "ordered"),
  exclude_chr = character(0)
)
}
\arguments{
\item{arrange_1L_chr}{Arrange (a character vector of length one), Default: c("default", "semi", "ordered")}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}
}
\value{
Disciplines (a character vector)
}
\description{
make_disciplines() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make disciplines. The function returns Disciplines (a character vector).
}
\keyword{internal}
