% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_draws_tb}
\alias{make_draws_tb}
\title{Make draws tibble}
\usage{
make_draws_tb(
  inputs_ls,
  iterations_int = 1:100,
  scale_1L_int = 1L,
  seed_1L_int = integer(0)
)
}
\arguments{
\item{inputs_ls}{Inputs (a list)}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100}

\item{scale_1L_int}{Scale (an integer vector of length one), Default: 1}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: integer(0)}
}
\value{
Draws (a tibble)
}
\description{
make_draws_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make draws tibble. The function returns Draws (a tibble).
}
\keyword{internal}
