% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_economic_results}
\alias{make_economic_results}
\title{Make economic results}
\usage{
make_economic_results(sim_results_ls, intervention_1L_chr, utilities_chr)
}
\arguments{
\item{sim_results_ls}{Sim results (a list)}

\item{intervention_1L_chr}{Intervention (a character vector of length one)}

\item{utilities_chr}{Utilities (a character vector)}
}
\value{
Economic results (a list)
}
\description{
make_economic_results() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make economic results. The function returns Economic results (a list).
}
