% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_economic_summary}
\alias{make_economic_summary}
\title{Make economic summary}
\usage{
make_economic_summary(
  sim_results_ls,
  correspondences_r3 = ready4show::ready4show_correspondences(),
  costs_1L_chr = "Cost",
  effects_1L_chr = "QALYs",
  reference_1L_chr = "Intervention",
  threshold_1L_dbl = 96000,
  what_1L_chr = "total"
)
}
\arguments{
\item{sim_results_ls}{Sim results (a list)}

\item{correspondences_r3}{Correspondences (a ready4 submodule), Default: ready4show::ready4show_correspondences()}

\item{costs_1L_chr}{Costs (a character vector of length one), Default: 'Cost'}

\item{effects_1L_chr}{Effects (a character vector of length one), Default: 'QALYs'}

\item{reference_1L_chr}{Reference (a character vector of length one), Default: 'Intervention'}

\item{threshold_1L_dbl}{Threshold (a double vector of length one), Default: 96000}

\item{what_1L_chr}{What (a character vector of length one), Default: 'total'}
}
\value{
Economic (an output object of multiple potential types)
}
\description{
make_economic_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make economic summary. The function returns Economic (an output object of multiple potential types).
}
\keyword{internal}
