% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_enhanced_pool}
\alias{make_enhanced_pool}
\title{Make enhanced pool}
\usage{
make_enhanced_pool(pooled_fits_ls, arguments_ls)
}
\arguments{
\item{pooled_fits_ls}{Pooled fits (a list)}

\item{arguments_ls}{Arguments (a list)}
}
\value{
Pooled (a list)
}
\description{
make_enhanced_pool() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make enhanced pool. The function returns Pooled (a list).
}
