% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_episodes_lup}
\alias{make_episodes_lup}
\title{Make episodes lookup table}
\usage{
make_episodes_lup(
  processed_ls,
  provider_lup_tb,
  add_programs_int = integer(0),
  distinct_orgs_1L_lgl = TRUE,
  drop_chr = character(0),
  filter_1L_lgl = TRUE,
  filter_fn = identity,
  keep_chr = c("client_key", "program_type", "referral_date"),
  phn_code_1L_chr = "PHN_code",
  phn_name_1L_chr = "PHN_area_name",
  program_services_lup = NULL,
  program_true_1L_chr = "is_program",
  program_true_int = integer(0),
  type_1L_chr = c("one", "two")
)
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{provider_lup_tb}{Provider lookup table (a tibble)}

\item{add_programs_int}{Add programs (an integer vector), Default: integer(0)}

\item{distinct_orgs_1L_lgl}{Distinct organisations (a logical vector of length one), Default: TRUE}

\item{drop_chr}{Drop (a character vector), Default: character(0)}

\item{filter_1L_lgl}{Filter (a logical vector of length one), Default: TRUE}

\item{filter_fn}{Filter (a function), Default: identity}

\item{keep_chr}{Keep (a character vector), Default: c("client_key", "program_type", "referral_date")}

\item{phn_code_1L_chr}{Primary Health Network code (a character vector of length one), Default: 'PHN_code'}

\item{phn_name_1L_chr}{Primary Health Network name (a character vector of length one), Default: 'PHN_area_name'}

\item{program_services_lup}{Program services (a lookup table), Default: NULL}

\item{program_true_1L_chr}{Program true (a character vector of length one), Default: 'is_program'}

\item{program_true_int}{Program true (an integer vector), Default: integer(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("one", "two")}
}
\value{
Episodes lookup table (a tibble)
}
\description{
make_episodes_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make episodes lookup table. The function returns Episodes lookup table (a tibble).
}
\keyword{internal}
