% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_experts_tb}
\alias{make_experts_tb}
\title{Make experts tibble}
\usage{
make_experts_tb(experts_ls, anonymise_1L_lgl = T, prefix_1L_chr = "Expert ")
}
\arguments{
\item{experts_ls}{Experts (a list)}

\item{anonymise_1L_lgl}{Anonymise (a logical vector of length one), Default: T}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Expert '}
}
\value{
Experts (a tibble)
}
\description{
make_experts_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make experts tibble. The function returns Experts (a tibble).
}
\keyword{internal}
