% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_iar_params}
\alias{make_iar_params}
\title{Make Initial Assessment andeferral parameters}
\usage{
make_iar_params(
  processed_ls,
  raw_mds_data_ls,
  test_1L_chr,
  comparator_1L_chr = "Comparator",
  comparator_int = integer(0),
  intervention_1L_chr = "Intervention",
  type_1L_chr = c(intervention_1L_chr, comparator_1L_chr),
  what_1L_chr = "InHouseIAR"
)
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{raw_mds_data_ls}{Raw Minimum Dataset data (a list)}

\item{test_1L_chr}{Test (a character vector of length one)}

\item{comparator_1L_chr}{Comparator (a character vector of length one), Default: 'Comparator'}

\item{comparator_int}{Comparator (an integer vector), Default: integer(0)}

\item{intervention_1L_chr}{Intervention (a character vector of length one), Default: 'Intervention'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c(intervention_1L_chr, comparator_1L_chr)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'InHouseIAR'}
}
\value{
Parameters (a double vector)
}
\description{
make_iar_params() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make initial assessment andeferral parameters. The function returns Parameters (a double vector).
}
\keyword{internal}
