% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ineligibility_fns_ls}
\alias{make_ineligibility_fns_ls}
\title{Make ineligibility functions list}
\usage{
make_ineligibility_fns_ls(post_fn = identity, pre_fn = identity)
}
\arguments{
\item{post_fn}{Post (a function), Default: identity}

\item{pre_fn}{Pre (a function), Default: identity}
}
\value{
Ineligibility (a list)
}
\description{
make_ineligibility_fns_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ineligibility functions list. The function returns Ineligibility (a list).
}
\keyword{internal}
