% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_initialise_ls}
\alias{make_initialise_ls}
\title{Make initialise list}
\usage{
make_initialise_ls(
  default_fn = identity,
  derive_ls = list(),
  update_fn = identity
)
}
\arguments{
\item{default_fn}{Default (a function), Default: identity}

\item{derive_ls}{Derive (a list), Default: list()}

\item{update_fn}{Update (a function), Default: identity}
}
\value{
Initialise (a list)
}
\description{
make_initialise_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make initialise list. The function returns Initialise (a list).
}
\keyword{internal}
