% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_interactions_summary}
\alias{make_interactions_summary}
\title{Make interactions summary}
\usage{
make_interactions_summary(processed_ls)
}
\arguments{
\item{processed_ls}{Processed (a list)}
}
\value{
Interactions (a tibble)
}
\description{
make_interactions_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make interactions summary. The function returns Interactions (a tibble).
}
