% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_joiners_series}
\alias{make_joiners_series}
\title{Make joiners series}
\usage{
make_joiners_series(model_data_ls)
}
\arguments{
\item{model_data_ls}{Model data (a list)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
make_joiners_series() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make joiners series. The function is called for its side effects and does not return a value.
}
