% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_k10_change_tb}
\alias{make_k10_change_tb}
\title{Make K10 change tibble}
\usage{
make_k10_change_tb(population_k10_tb)
}
\arguments{
\item{population_k10_tb}{Population K10 (a tibble)}
}
\value{
K10 change (a tibble)
}
\description{
make_k10_change_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make k10 change tibble. The function returns K10 change (a tibble).
}
