% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mapping_params_tb}
\alias{make_mapping_params_tb}
\title{Make mapping parameters tibble}
\usage{
make_mapping_params_tb()
}
\value{
Mapping parameters (a tibble)
}
\description{
make_mapping_params_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make mapping parameters tibble. The function returns Mapping parameters (a tibble).
}
\keyword{internal}
