% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_clients_tb}
\alias{make_mds_clients_tb}
\title{Make Minimum Dataset clients tibble}
\usage{
make_mds_clients_tb(raw_mds_data_ls)
}
\arguments{
\item{raw_mds_data_ls}{Raw Minimum Dataset data (a list)}
}
\value{
a Minimum Dataset clients (a tibble)
}
\description{
make_mds_clients_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset clients tibble. The function returns a Minimum Dataset clients (a tibble).
}
