% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_collection_tb}
\alias{make_mds_collection_tb}
\title{Make Minimum Dataset collection tibble}
\usage{
make_mds_collection_tb(
  raw_mds_data_ls,
  drop_chr = c("collection_occasion_tags", "X.covid19")
)
}
\arguments{
\item{raw_mds_data_ls}{Raw Minimum Dataset data (a list)}

\item{drop_chr}{Drop (a character vector), Default: c("collection_occasion_tags", "X.covid19")}
}
\value{
Collection (a tibble)
}
\description{
make_mds_collection_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset collection tibble. The function returns Collection (a tibble).
}
