% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_episodes_tb}
\alias{make_mds_episodes_tb}
\title{Make Minimum Dataset episodes tibble}
\usage{
make_mds_episodes_tb(raw_mds_data_ls, processed_ls)
}
\arguments{
\item{raw_mds_data_ls}{Raw Minimum Dataset data (a list)}

\item{processed_ls}{Processed (a list)}
}
\value{
Episodes (a tibble)
}
\description{
make_mds_episodes_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset episodes tibble. The function returns Episodes (a tibble).
}
