% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_processed_ls}
\alias{make_mds_processed_ls}
\title{Make Minimum Dataset processed list}
\usage{
make_mds_processed_ls(
  raw_mds_data_ls,
  add_start_date_1L_lgl,
  program_services_lup,
  program_true_chr,
  program_type_ls,
  provider_lup_tb,
  filter_fn = identity,
  frequencies_chr = c(paste0("k10p_item", c(1:10, 14))),
  integers_chr = c("k10p_score", paste0("k10p_item", 11:13),
    "iar_dst_practitioner_level_of_care"),
  mature_after_dtm = lubridate::years(1),
  outcomes_ls = make_project_2_outcomes_ls(),
  review_target_dtm = lubridate::days(90)
)
}
\arguments{
\item{raw_mds_data_ls}{Raw Minimum Dataset data (a list)}

\item{add_start_date_1L_lgl}{Add start date (a logical vector of length one)}

\item{program_services_lup}{Program services (a lookup table)}

\item{program_true_chr}{Program true (a character vector)}

\item{program_type_ls}{Program type (a list)}

\item{provider_lup_tb}{Provider lookup table (a tibble)}

\item{filter_fn}{Filter (a function), Default: identity}

\item{frequencies_chr}{Frequencies (a character vector), Default: c(paste0("k10p_item", c(1:10, 14)))}

\item{integers_chr}{Integers (a character vector), Default: c("k10p_score", paste0("k10p_item", 11:13), "iar_dst_practitioner_level_of_care")}

\item{mature_after_dtm}{Mature after (a date vector), Default: lubridate::years(1)}

\item{outcomes_ls}{Outcomes (a list), Default: make_project_2_outcomes_ls()}

\item{review_target_dtm}{Review target (a date vector), Default: lubridate::days(90)}
}
\value{
Processed (a list)
}
\description{
make_mds_processed_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset processed list. The function returns Processed (a list).
}
