% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_program_lup}
\alias{make_mds_program_lup}
\title{Make Minimum Dataset program lookup table}
\usage{
make_mds_program_lup()
}
\value{
Program (a lookup table)
}
\description{
make_mds_program_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset program lookup table. The function returns Program (a lookup table).
}
