% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_program_starts}
\alias{make_mds_program_starts}
\title{Make Minimum Dataset program starts}
\usage{
make_mds_program_starts(
  processed_ls,
  filter_fn,
  program_services_lup,
  program_true_chr,
  program_type_ls,
  provider_lup_tb,
  add_start_date_1L_lgl = FALSE,
  mature_after_dtm = lubridate::years(1)
)
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{filter_fn}{Filter (a function)}

\item{program_services_lup}{Program services (a lookup table)}

\item{program_true_chr}{Program true (a character vector)}

\item{program_type_ls}{Program type (a list)}

\item{provider_lup_tb}{Provider lookup table (a tibble)}

\item{add_start_date_1L_lgl}{Add start date (a logical vector of length one), Default: FALSE}

\item{mature_after_dtm}{Mature after (a date vector), Default: lubridate::years(1)}
}
\value{
Start up (a lookup table)
}
\description{
make_mds_program_starts() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset program starts. The function returns Start up (a lookup table).
}
