% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_providers_tb}
\alias{make_mds_providers_tb}
\title{Make Minimum Dataset providers tibble}
\usage{
make_mds_providers_tb(raw_mds_data_ls, additional_chr = character(0))
}
\arguments{
\item{raw_mds_data_ls}{Raw Minimum Dataset data (a list)}

\item{additional_chr}{Additional (a character vector), Default: character(0)}
}
\value{
Providers (a tibble)
}
\description{
make_mds_providers_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset providers tibble. The function returns Providers (a tibble).
}
\keyword{internal}
