% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_service_contacts}
\alias{make_mds_service_contacts}
\title{Make Minimum Dataset service contacts}
\usage{
make_mds_service_contacts(raw_mds_data_ls)
}
\arguments{
\item{raw_mds_data_ls}{Raw Minimum Dataset data (a list)}
}
\value{
Data (a tibble)
}
\description{
make_mds_service_contacts() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset service contacts. The function returns Data (a tibble).
}
\keyword{internal}
