% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_services_tb}
\alias{make_mds_services_tb}
\title{Make Minimum Dataset services tibble}
\usage{
make_mds_services_tb(
  processed_ls,
  end_dtm = lubridate::NA_Date_,
  episode_keys_chr = character(0),
  start_dtm = lubridate::NA_Date_,
  summarise_1L_lgl = FALSE
)
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{end_dtm}{End (a date vector), Default: lubridate::NA_Date_}

\item{episode_keys_chr}{Episode keys (a character vector), Default: character(0)}

\item{start_dtm}{Start (a date vector), Default: lubridate::NA_Date_}

\item{summarise_1L_lgl}{Summarise (a logical vector of length one), Default: FALSE}
}
\value{
Services (a tibble)
}
\description{
make_mds_services_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset services tibble. The function returns Services (a tibble).
}
