% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mds_unique_measures}
\alias{make_mds_unique_measures}
\title{Make Minimum Dataset unique measures}
\usage{
make_mds_unique_measures(
  data_tb,
  episodes_tb,
  review_target_dtm = lubridate::days(90),
  prefixes_chr = c("k10p_", "iar_dst_")
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{episodes_tb}{Episodes (a tibble)}

\item{review_target_dtm}{Review target (a date vector), Default: lubridate::days(90)}

\item{prefixes_chr}{Prefixes (a character vector), Default: c("k10p_", "iar_dst_")}
}
\value{
Data (a tibble)
}
\description{
make_mds_unique_measures() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make minimum dataset unique measures. The function returns Data (a tibble).
}
\keyword{internal}
