% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_model_dyad_ls}
\alias{make_model_dyad_ls}
\title{Make model dyad list}
\usage{
make_model_dyad_ls(
  X_Ready4useDyad = ready4use::Ready4useDyad(),
  Y_Ready4useDyad = ready4use::Ready4useDyad()
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}
}
\value{
Model dyad (a list)
}
\description{
make_model_dyad_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model dyad list. The function returns Model dyad (a list).
}
