% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_outcomes_tb}
\alias{make_outcomes_tb}
\title{Make outcomes tibble}
\usage{
make_outcomes_tb(
  outcomes_chr = character(0),
  categories_chr = character(0),
  derivation_chr = character(0),
  subcategories_chr = character(0)
)
}
\arguments{
\item{outcomes_chr}{Outcomes (a character vector), Default: character(0)}

\item{categories_chr}{Categories (a character vector), Default: character(0)}

\item{derivation_chr}{Derivation (a character vector), Default: character(0)}

\item{subcategories_chr}{Subcategories (a character vector), Default: character(0)}
}
\value{
Outcomes (a tibble)
}
\description{
make_outcomes_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make outcomes tibble. The function returns Outcomes (a tibble).
}
\keyword{internal}
