% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_outcomes_vars}
\alias{make_outcomes_vars}
\title{Make outcomes variables}
\usage{
make_outcomes_vars(
  X_Ready4useDyad,
  Y_Ready4useDyad,
  Z_Ready4useDyad,
  exclude_chr = character(0),
  exclude_suffixes_chr = character(0),
  modifiable_chr = character(0),
  numeric_only_1L_lgl = FALSE
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Z_Ready4useDyad}{PARAM_DESCRIPTION}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{exclude_suffixes_chr}{Exclude suffixes (a character vector), Default: character(0)}

\item{modifiable_chr}{Modifiable (a character vector), Default: character(0)}

\item{numeric_only_1L_lgl}{Numeric only (a logical vector of length one), Default: FALSE}
}
\value{
Outcomes (a character vector)
}
\description{
make_outcomes_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make outcomes variables. The function returns Outcomes (a character vector).
}
\keyword{internal}
