% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_parsnip_mdl}
\alias{make_parsnip_mdl}
\title{Make parsnip model}
\usage{
make_parsnip_mdl(
  data_tb,
  model_fn = parsnip::multinom_reg,
  model_args_ls = NULL,
  x_chr,
  y_1L_chr,
  ...
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{model_fn}{Model (a function), Default: parsnip::multinom_reg}

\item{model_args_ls}{Model arguments (a list), Default: NULL}

\item{x_chr}{X (a character vector)}

\item{y_1L_chr}{Y (a character vector of length one)}

\item{...}{Additional arguments}
}
\value{
Model (a model)
}
\description{
make_parsnip_mdl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make parsnip model. The function returns Model (a model).
}
