% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_phn_lup}
\alias{make_phn_lup}
\title{Make Primary Health Network lookup table}
\usage{
make_phn_lup(
  code_1L_chr = character(0),
  name_1L_chr = character(0),
  jurisdiction_1L_chr = character(0)
)
}
\arguments{
\item{code_1L_chr}{Code (a character vector of length one), Default: character(0)}

\item{name_1L_chr}{Name (a character vector of length one), Default: character(0)}

\item{jurisdiction_1L_chr}{Jurisdiction (a character vector of length one), Default: character(0)}
}
\value{
a Primary Health Network lookup table (an output object of multiple potential types)
}
\description{
make_phn_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make primary health network lookup table. The function returns a Primary Health Network lookup table (an output object of multiple potential types).
}
