% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pooled_fit}
\alias{make_pooled_fit}
\title{Make pooled fit}
\usage{
make_pooled_fit(experts_tb, question_1L_chr = character(0))
}
\arguments{
\item{experts_tb}{Experts (a tibble)}

\item{question_1L_chr}{Question (a character vector of length one), Default: character(0)}
}
\value{
Fit (an output object of multiple potential types)
}
\description{
make_pooled_fit() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make pooled fit. The function returns Fit (an output object of multiple potential types).
}
