% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_postcode_lup}
\alias{make_postcode_lup}
\title{Make postcode lookup table}
\usage{
make_postcode_lup(url_1L_chr = character(0))
}
\arguments{
\item{url_1L_chr}{Url (a character vector of length one), Default: character(0)}
}
\value{
Postcode (a lookup table)
}
\description{
make_postcode_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make postcode lookup table. The function returns Postcode (a lookup table).
}
