% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_append_ls}
\alias{make_project_2_append_ls}
\title{Make project 2 append list}
\usage{
make_project_2_append_ls(
  arms_for_iar_adjustment_chr,
  arms_for_non_helpseeking_chr,
  never_1L_int
)
}
\arguments{
\item{arms_for_iar_adjustment_chr}{Arms for Initial Assessment andeferral adjustment (a character vector)}

\item{arms_for_non_helpseeking_chr}{Arms for non helpseeking (a character vector)}

\item{never_1L_int}{Never (an integer vector of length one)}
}
\value{
Append (a list)
}
\description{
make_project_2_append_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 append list. The function returns Append (a list).
}
\keyword{internal}
