% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_arms_extras_ls}
\alias{make_project_2_arms_extras_ls}
\title{Make project 2 arms extras list}
\usage{
make_project_2_arms_extras_ls(
  arms_chr,
  arms_for_iar_adjustment_chr = character(0),
  arms_for_intervention_costs_chr = character(0),
  arms_for_non_helpseeking_chr = character(0),
  arms_for_offsets_chr = character(0),
  treatment_ls = NULL,
  tx_duration_dtm = lubridate::weeks(12)
)
}
\arguments{
\item{arms_chr}{Arms (a character vector)}

\item{arms_for_iar_adjustment_chr}{Arms for Initial Assessment andeferral adjustment (a character vector), Default: character(0)}

\item{arms_for_intervention_costs_chr}{Arms for intervention costs (a character vector), Default: character(0)}

\item{arms_for_non_helpseeking_chr}{Arms for non helpseeking (a character vector), Default: character(0)}

\item{arms_for_offsets_chr}{Arms for offsets (a character vector), Default: character(0)}

\item{treatment_ls}{Treatment (a list), Default: NULL}

\item{tx_duration_dtm}{Treatment duration (a date vector), Default: lubridate::weeks(12)}
}
\value{
Arms extras (a list)
}
\description{
make_project_2_arms_extras_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 arms extras list. The function returns Arms extras (a list).
}
\keyword{internal}
