% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_derive_ls}
\alias{make_project_2_derive_ls}
\title{Make project 2 derive list}
\usage{
make_project_2_derive_ls(
  function_fn = NULL,
  discard_chr = character(0),
  keep_chr = character(0)
)
}
\arguments{
\item{function_fn}{Function (a function), Default: NULL}

\item{discard_chr}{Discard (a character vector), Default: character(0)}

\item{keep_chr}{Keep (a character vector), Default: character(0)}
}
\value{
Derive (a list)
}
\description{
make_project_2_derive_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 derive list. The function returns Derive (a list).
}
\keyword{internal}
