% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_initialise_ls}
\alias{make_project_2_initialise_ls}
\title{Make project 2 initialise list}
\usage{
make_project_2_initialise_ls(derive_ls = list())
}
\arguments{
\item{derive_ls}{Derive (a list), Default: list()}
}
\value{
Initialise (a list)
}
\description{
make_project_2_initialise_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 initialise list. The function returns Initialise (a list).
}
\keyword{internal}
