% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_k10_change_mdls}
\alias{make_project_2_k10_change_mdls}
\title{Make project 2 K10 change models}
\usage{
make_project_2_k10_change_mdls(
  X_Ready4useDyad,
  type_1L_chr = c("Intervention", "Jurisdiction")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("Intervention", "Jurisdiction")}
}
\value{
K10 (a list)
}
\description{
make_project_2_k10_change_mdls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 k10 change models. The function returns K10 (a list).
}
\keyword{internal}
