% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_k10_relapse_mdls}
\alias{make_project_2_k10_relapse_mdls}
\title{Make project 2 K10 relapse models}
\usage{
make_project_2_k10_relapse_mdls(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
K10 (a list)
}
\description{
make_project_2_k10_relapse_mdls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 k10 relapse models. The function returns K10 (a list).
}
\keyword{internal}
