% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_labels}
\alias{make_project_2_labels}
\title{Make project 2 labels}
\usage{
make_project_2_labels(type_1L_chr = c("regression", "simulation"))
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one), Default: c("regression", "simulation")}
}
\value{
Labels (a list)
}
\description{
make_project_2_labels() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 labels. The function returns Labels (a list).
}
\keyword{internal}
