% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_minutes_mdls}
\alias{make_project_2_minutes_mdls}
\title{Make project 2 minutes models}
\usage{
make_project_2_minutes_mdls(
  X_Ready4useDyad,
  disciplines_chr = make_disciplines(),
  family_2_1L_chr = "Gamma(link = 'log')",
  link_1_1L_chr = "logit",
  x_part_1_ls = NULL,
  x_part_2_ls = NULL,
  y_1L_chr = "TotalUseMins_change",
  use_1_int = integer(0),
  use_2_int = integer(0),
  ...
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{disciplines_chr}{Disciplines (a character vector), Default: make_disciplines()}

\item{family_2_1L_chr}{Family 2 (a character vector of length one), Default: 'Gamma(link = 'log')'}

\item{link_1_1L_chr}{Link 1 (a character vector of length one), Default: 'logit'}

\item{x_part_1_ls}{X part 1 (a list), Default: NULL}

\item{x_part_2_ls}{X part 2 (a list), Default: NULL}

\item{y_1L_chr}{Y (a character vector of length one), Default: 'TotalUseMins_change'}

\item{use_1_int}{Use 1 (an integer vector), Default: integer(0)}

\item{use_2_int}{Use 2 (an integer vector), Default: integer(0)}

\item{...}{Additional arguments}
}
\value{
Tpm models (a list)
}
\description{
make_project_2_minutes_mdls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 minutes models. The function returns Tpm models (a list).
}
