% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_outcomes_ls}
\alias{make_project_2_outcomes_ls}
\title{Make project 2 outcomes list}
\usage{
make_project_2_outcomes_ls()
}
\value{
Outcomes (a list)
}
\description{
make_project_2_outcomes_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 outcomes list. The function returns Outcomes (a list).
}
\keyword{internal}
