% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_outcomes_tb}
\alias{make_project_2_outcomes_tb}
\title{Make project 2 outcomes tibble}
\usage{
make_project_2_outcomes_tb()
}
\value{
Outcomes (a tibble)
}
\description{
make_project_2_outcomes_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 outcomes tibble. The function returns Outcomes (a tibble).
}
\keyword{internal}
