% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_regression_to_mean}
\alias{make_project_2_regression_to_mean}
\title{Make project 2 regression to mean}
\usage{
make_project_2_regression_to_mean(
  event_nm_1L_chr = "RegressionToMean",
  functions_ls = make_ineligibility_fns_ls(),
  ineligible_1L_chr = character(0),
  outcome_var_1L_chr = "K10",
  use_schedule_1L_chr = "Y",
  use_trigger_1L_chr = "Z"
)
}
\arguments{
\item{event_nm_1L_chr}{Event name (a character vector of length one), Default: 'RegressionToMean'}

\item{functions_ls}{Functions (a list), Default: make_ineligibility_fns_ls()}

\item{ineligible_1L_chr}{Ineligible (a character vector of length one), Default: character(0)}

\item{outcome_var_1L_chr}{Outcome variable (a character vector of length one), Default: 'K10'}

\item{use_schedule_1L_chr}{Use schedule (a character vector of length one), Default: 'Y'}

\item{use_trigger_1L_chr}{Use trigger (a character vector of length one), Default: 'Z'}
}
\value{
X (Model event scheduling and event logic data.)
}
\description{
make_project_2_regression_to_mean() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 regression to mean. The function is called for its side effects and does not return a value.
}
\keyword{internal}
