% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_report}
\alias{make_project_2_report}
\title{Make project 2 report}
\usage{
make_project_2_report(
  model_data_ls = NULL,
  arms_chr,
  params_tb,
  period_dtm = lubridate::years(1),
  platform_1L_chr = "Intervention",
  processed_ls = NULL,
  regressions_ls = NULL,
  select_1L_chr = character(0),
  sim_results_ls = NULL,
  timepoint_1L_chr = character(0),
  transformations_chr = character(0),
  type_1L_chr = "full_combos",
  ungroup_1L_lgl = FALSE,
  weeks_int = integer(0),
  what_1L_chr = c("paramscost", "paramsk10", "resultsaqol", "resultseq5d",
    "resultseconomic", "resultsoutcomes", "resultsutility", "serviceuse",
    "serviceusecost")
)
}
\arguments{
\item{model_data_ls}{Model data (a list), Default: NULL}

\item{arms_chr}{Arms (a character vector)}

\item{params_tb}{Parameters (a tibble)}

\item{period_dtm}{Period (a date vector), Default: lubridate::years(1)}

\item{platform_1L_chr}{Platform (a character vector of length one), Default: 'Intervention'}

\item{processed_ls}{Processed (a list), Default: NULL}

\item{regressions_ls}{Regressions (a list), Default: NULL}

\item{select_1L_chr}{Select (a character vector of length one), Default: character(0)}

\item{sim_results_ls}{Sim results (a list), Default: NULL}

\item{timepoint_1L_chr}{Timepoint (a character vector of length one), Default: character(0)}

\item{transformations_chr}{Transformations (a character vector), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'full_combos'}

\item{ungroup_1L_lgl}{Ungroup (a logical vector of length one), Default: FALSE}

\item{weeks_int}{Weeks (an integer vector), Default: integer(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("paramscost", "paramsk10", "resultsaqol", "resultseq5d", "resultseconomic",
"resultsoutcomes", "resultsutility", "serviceuse", "serviceusecost")}
}
\value{
Data (an output object of multiple potential types)
}
\description{
make_project_2_report() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 report. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
