% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_resources_tb}
\alias{make_project_2_resources_tb}
\title{Make project 2 resources tibble}
\usage{
make_project_2_resources_tb()
}
\value{
Resources (a tibble)
}
\description{
make_project_2_resources_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 resources tibble. The function returns Resources (a tibble).
}
\keyword{internal}
