% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_sim_summary}
\alias{make_project_2_sim_summary}
\title{Make project 2 sim summary}
\usage{
make_project_2_sim_summary(
  sim_results_ls,
  arms_chr,
  element_1L_chr = "Z",
  groupings_chr = c("Diagnosis", "Distress"),
  order_1L_lgl = TRUE,
  convert_1L_lgl = TRUE,
  platform_1L_chr = "Intervention",
  select_1L_chr = "all",
  type_1L_chr = c("outcomes", "economic"),
  what_1L_chr = c("total", "diagnosis", "iar", "full_combos")
)
}
\arguments{
\item{sim_results_ls}{Sim results (a list)}

\item{arms_chr}{Arms (a character vector)}

\item{element_1L_chr}{Element (a character vector of length one), Default: 'Z'}

\item{groupings_chr}{Groupings (a character vector), Default: c("Diagnosis", "Distress")}

\item{order_1L_lgl}{Order (a logical vector of length one), Default: TRUE}

\item{convert_1L_lgl}{Convert (a logical vector of length one), Default: TRUE}

\item{platform_1L_chr}{Platform (a character vector of length one), Default: 'Intervention'}

\item{select_1L_chr}{Select (a character vector of length one), Default: 'all'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("outcomes", "economic")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("total", "diagnosis", "iar", "full_combos")}
}
\value{
Summary (a tibble)
}
\description{
make_project_2_sim_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 sim summary. The function returns Summary (a tibble).
}
