% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_synthetic_pop}
\alias{make_project_2_synthetic_pop}
\title{Make project 2 synthetic pop}
\usage{
make_project_2_synthetic_pop(model_data_ls, size_1L_int = 1000L)
}
\arguments{
\item{model_data_ls}{Model data (a list)}

\item{size_1L_int}{Size (an integer vector of length one), Default: 1000}
}
\value{
Population (a list)
}
\description{
make_project_2_synthetic_pop() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 synthetic pop. The function returns Population (a list).
}
