% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_tfmn_ls}
\alias{make_project_2_tfmn_ls}
\title{Make project 2 transformation list}
\usage{
make_project_2_tfmn_ls(type_1L_chr = "outcome")
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one), Default: 'outcome'}
}
\value{
Transformation (a list)
}
\description{
make_project_2_tfmn_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 transformation list. The function returns Transformation (a list).
}
\keyword{internal}
