% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_theme_fn}
\alias{make_project_2_theme_fn}
\title{Make project 2 theme function}
\usage{
make_project_2_theme_fn(output_1L_chr = c("Word", "PDF", "HTML"))
}
\arguments{
\item{output_1L_chr}{Output (a character vector of length one), Default: c("Word", "PDF", "HTML")}
}
\value{
Plot (a function)
}
\description{
make_project_2_theme_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 theme function. The function returns Plot (a function).
}
\keyword{internal}
