% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_aqol6d_mdls}
\alias{make_project_aqol6d_mdls}
\title{Make project Assessment of Quality of Life Six Dimension models}
\usage{
make_project_aqol6d_mdls(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
Assessment of Quality of Life Six Dimension (a list)
}
\description{
make_project_aqol6d_mdls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project assessment of quality of life six dimension models. The function returns Assessment of Quality of Life Six Dimension (a list).
}
\keyword{internal}
