% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_chu9d_mdls}
\alias{make_project_chu9d_mdls}
\title{Make project CHU-9D models}
\usage{
make_project_chu9d_mdls(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
a CHU-9D (a list)
}
\description{
make_project_chu9d_mdls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project chu-9d models. The function returns a CHU-9D (a list).
}
