% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_cmprsn}
\alias{make_project_cmprsn}
\title{Make project comparison}
\usage{
make_project_cmprsn(
  model_data_ls,
  by_1L_chr,
  include_chr,
  what_1L_chr = "X",
  digits_1L_int = 2L,
  labels_ls = NULL,
  tfmn_fn = identity,
  tfmn_1_fn = identity,
  tfmn_2_fn = identity,
  type_1L_chr = c("unimputed", "imputed", "combined")
)
}
\arguments{
\item{model_data_ls}{Model data (a list)}

\item{by_1L_chr}{By (a character vector of length one)}

\item{include_chr}{Include (a character vector)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'X'}

\item{digits_1L_int}{Digits (an integer vector of length one), Default: 2}

\item{labels_ls}{Labels (a list), Default: NULL}

\item{tfmn_fn}{Transformation (a function), Default: identity}

\item{tfmn_1_fn}{Transformation 1 (a function), Default: identity}

\item{tfmn_2_fn}{Transformation 2 (a function), Default: identity}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("unimputed", "imputed", "combined")}
}
\value{
Comparison (an output object of multiple potential types)
}
\description{
make_project_cmprsn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project comparison. The function returns Comparison (an output object of multiple potential types).
}
