% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_consolidated_ds}
\alias{make_project_consolidated_ds}
\title{Make project consolidated dataset}
\usage{
make_project_consolidated_ds(
  X_Ready4useDyad,
  Y_Ready4useDyad,
  Z_Ready4useDyad = ready4use::Ready4useDyad,
  type_1L_chr = c("outcomes", "tx_status")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Z_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("outcomes", "tx_status")}
}
\value{
A (A dataset and data dictionary pair.)
}
\description{
make_project_consolidated_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project consolidated dataset. The function is called for its side effects and does not return a value.
}
