% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_contacts_ds}
\alias{make_project_contacts_ds}
\title{Make project contacts dataset}
\usage{
make_project_contacts_ds(
  raw_data_ls,
  demographics_tb,
  recode_lup_r3 = make_project_recode_lup()
)
}
\arguments{
\item{raw_data_ls}{Raw data (a list)}

\item{demographics_tb}{Demographics (a tibble)}

\item{recode_lup_r3}{Recode (a ready4 submodule extension of lookup table), Default: make_project_recode_lup()}
}
\value{
Contacts (a tibble)
}
\description{
make_project_contacts_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project contacts dataset. The function returns Contacts (a tibble).
}
\keyword{internal}
