% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_cost_mdlng_ds}
\alias{make_project_cost_mdlng_ds}
\title{Make project cost modelling dataset}
\usage{
make_project_cost_mdlng_ds(
  W_Ready4useDyad,
  X_Ready4useDyad,
  transform_gender_1L_lgl = T
)
}
\arguments{
\item{W_Ready4useDyad}{PARAM_DESCRIPTION}

\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{transform_gender_1L_lgl}{Transform gender (a logical vector of length one), Default: T}
}
\value{
Z (A dataset and data dictionary pair.)
}
\description{
make_project_cost_mdlng_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project cost modelling dataset. The function is called for its side effects and does not return a value.
}
