% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_costs_ds}
\alias{make_project_costs_ds}
\title{Make project costs dataset}
\usage{
make_project_costs_ds(
  dss_ls,
  end_dtm = NULL,
  financial_yrs_chr = c("FY 2021", "FY 2022", "FY 2023", "FY 2024"),
  price_indices_dbl = c(97.2, 100, 100.7796, 102.5514),
  ref_1L_chr = "FY 2024",
  start_dtm = NULL,
  sunk_ls = list(Base = make_project_sunk_tb(), S1 = make_project_sunk_tb(0)),
  type_1L_chr = c("constant", "current"),
  what_1L_chr = c("all", "fixed", "variable", "unit", "initial")
)
}
\arguments{
\item{dss_ls}{Datasets (a list)}

\item{end_dtm}{End (a date vector), Default: NULL}

\item{financial_yrs_chr}{Financial years (a character vector), Default: c("FY 2021", "FY 2022", "FY 2023", "FY 2024")}

\item{price_indices_dbl}{Price indices (a double vector), Default: c(97.2, 100, 100.7796, 102.5514)}

\item{ref_1L_chr}{Reference (a character vector of length one), Default: 'FY 2024'}

\item{start_dtm}{Start (a date vector), Default: NULL}

\item{sunk_ls}{Sunk (a list), Default: list(Base = make_project_sunk_tb(), S1 = make_project_sunk_tb(0))}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("constant", "current")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("all", "fixed", "variable", "unit", "initial")}
}
\value{
Costs (a tibble)
}
\description{
make_project_costs_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project costs dataset. The function returns Costs (a tibble).
}
\keyword{internal}
