% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_joiners_ds}
\alias{make_project_joiners_ds}
\title{Make project joiners dataset}
\usage{
make_project_joiners_ds(processed_ls, as_dyad_1L_lgl = T)
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{as_dyad_1L_lgl}{As dyad (a logical vector of length one), Default: T}
}
\value{
Data (an output object of multiple potential types)
}
\description{
make_project_joiners_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project joiners dataset. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
