% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_keys}
\alias{make_project_keys}
\title{Make project keys}
\usage{
make_project_keys(type_1L_chr = c("micro", "ts"))
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one), Default: c("micro", "ts")}
}
\value{
Keys (a character vector)
}
\description{
make_project_keys() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project keys. The function returns Keys (a character vector).
}
\keyword{internal}
