% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_metrics}
\alias{make_project_metrics}
\title{Make project metrics}
\usage{
make_project_metrics(unit_costs_tb)
}
\arguments{
\item{unit_costs_tb}{Unit costs (a tibble)}
}
\value{
Metrics (a character vector)
}
\description{
make_project_metrics() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project metrics. The function returns Metrics (a character vector).
}
