% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_minutes_mdls}
\alias{make_project_minutes_mdls}
\title{Make project minutes models}
\usage{
make_project_minutes_mdls(
  X_Ready4useDyad,
  family_2_1L_chr = "Gamma(link = 'inverse')",
  link_1_1L_chr = "logit",
  x_part_1_ls = NULL,
  x_part_2_ls = NULL,
  y_1L_chr = "Minutes_change",
  ...
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{family_2_1L_chr}{Family 2 (a character vector of length one), Default: 'Gamma(link = 'inverse')'}

\item{link_1_1L_chr}{Link 1 (a character vector of length one), Default: 'logit'}

\item{x_part_1_ls}{X part 1 (a list), Default: NULL}

\item{x_part_2_ls}{X part 2 (a list), Default: NULL}

\item{y_1L_chr}{Y (a character vector of length one), Default: 'Minutes_change'}

\item{...}{Additional arguments}
}
\value{
Tpm models (a list)
}
\description{
make_project_minutes_mdls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project minutes models. The function returns Tpm models (a list).
}
