% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_notes_ds}
\alias{make_project_notes_ds}
\title{Make project notes dataset}
\usage{
make_project_notes_ds(raw_data_ls, financial_yrs_chr)
}
\arguments{
\item{raw_data_ls}{Raw data (a list)}

\item{financial_yrs_chr}{Financial years (a character vector)}
}
\value{
Notes (a tibble)
}
\description{
make_project_notes_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project notes dataset. The function returns Notes (a tibble).
}
\keyword{internal}
