% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_overview_ds}
\alias{make_project_overview_ds}
\title{Make project overview dataset}
\usage{
make_project_overview_ds(raw_data_ls, demographics_tb)
}
\arguments{
\item{raw_data_ls}{Raw data (a list)}

\item{demographics_tb}{Demographics (a tibble)}
}
\value{
Overview (a tibble)
}
\description{
make_project_overview_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project overview dataset. The function returns Overview (a tibble).
}
\keyword{internal}
