% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_params_tb}
\alias{make_project_params_tb}
\title{Make project parameters tibble}
\usage{
make_project_params_tb()
}
\value{
Parameters (a tibble)
}
\description{
make_project_params_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project parameters tibble. The function returns Parameters (a tibble).
}
\keyword{internal}
