% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_recode_lup}
\alias{make_project_recode_lup}
\title{Make project recode lookup table}
\usage{
make_project_recode_lup()
}
\value{
Recode (a ready4 submodule extension of lookup table)
}
\description{
make_project_recode_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project recode lookup table. The function returns Recode (a ready4 submodule extension of lookup table).
}
\keyword{internal}
