% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_service_use_ds}
\alias{make_project_service_use_ds}
\title{Make project service use dataset}
\usage{
make_project_service_use_ds(
  processed_ls,
  data_extract_dtm = as.POSIXct("2024-10-25")
)
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{data_extract_dtm}{Data extract (a date vector), Default: as.POSIXct("2024-10-25")}
}
\value{
Y (A dataset and data dictionary pair.)
}
\description{
make_project_service_use_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project service use dataset. The function is called for its side effects and does not return a value.
}
