% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_sunk_tb}
\alias{make_project_sunk_tb}
\title{Make project sunk tibble}
\usage{
make_project_sunk_tb(
  global_1L_dbl = numeric(0),
  content_dbl = 0.9,
  development_dbl = 0.75,
  implementation_dbl = 0.25,
  infrastructure_dbl = 0,
  management_dbl = 0,
  marketing_dbl = 0,
  onboarding_dbl = 0.1,
  reporting_dbl = 0.9
)
}
\arguments{
\item{global_1L_dbl}{Global (a double vector of length one), Default: numeric(0)}

\item{content_dbl}{Content (a double vector), Default: 0.9}

\item{development_dbl}{Development (a double vector), Default: 0.75}

\item{implementation_dbl}{Implementation (a double vector), Default: 0.25}

\item{infrastructure_dbl}{Infrastructure (a double vector), Default: 0}

\item{management_dbl}{Management (a double vector), Default: 0}

\item{marketing_dbl}{Marketing (a double vector), Default: 0}

\item{onboarding_dbl}{Onboarding (a double vector), Default: 0.1}

\item{reporting_dbl}{Reporting (a double vector), Default: 0.9}
}
\value{
Sunk (a tibble)
}
\description{
make_project_sunk_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project sunk tibble. The function returns Sunk (a tibble).
}
