% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_test_ds}
\alias{make_project_test_ds}
\title{Make project test dataset}
\usage{
make_project_test_ds(model_data_ls, update_qalys_fn = identity)
}
\arguments{
\item{model_data_ls}{Model data (a list)}

\item{update_qalys_fn}{Update Quality Adjusted Life Years (a function), Default: identity}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
make_project_test_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project test dataset. The function is called for its side effects and does not return a value.
}
