% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_tx_mdlng_ds}
\alias{make_project_tx_mdlng_ds}
\title{Make project treatment modelling dataset}
\usage{
make_project_tx_mdlng_ds(
  X_Ready4useDyad,
  Y_Ready4useDyad,
  Z_Ready4useDyad,
  periods_chr = c(x = "-2 to 0 Weeks", y = "0 to 12 Weeks", z = "12 to 24 Weeks"),
  periods_ls = list(x = c(-2, 0), y = c(0, 12), z = c(12, 24)),
  treatment_vals_chr = c("Waitlist", "Treatment", "Discharged"),
  treatment_vars_chr = c("treatment_status", "treatment_status_t2")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Z_Ready4useDyad}{PARAM_DESCRIPTION}

\item{periods_chr}{Periods (a character vector), Default: c(x = "-2 to 0 Weeks", y = "0 to 12 Weeks", z = "12 to 24 Weeks")}

\item{periods_ls}{Periods (a list), Default: list(x = c(-2, 0), y = c(0, 12), z = c(12, 24))}

\item{treatment_vals_chr}{Treatment values (a character vector), Default: c("Waitlist", "Treatment", "Discharged")}

\item{treatment_vars_chr}{Treatment variables (a character vector), Default: c("treatment_status", "treatment_status_t2")}
}
\value{
A (A dataset and data dictionary pair.)
}
\description{
make_project_tx_mdlng_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project treatment modelling dataset. The function is called for its side effects and does not return a value.
}
