% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_report_data}
\alias{make_report_data}
\title{Make report data}
\usage{
make_report_data(
  model_data_ls = NULL,
  platform_1L_chr = "Intervention",
  processed_ls = NULL,
  regressions_ls = NULL,
  sim_results_ls = NULL,
  transformations_chr = character(0),
  what_1L_chr = c("descriptives", "costadj", "costitem", "costsum", "costunit",
    "minutes", "mnl-wait", "mnl-tx", "mnl-disc", "outcomes", "outcomeslong",
    "paramscost", "paramsk10", "resultsaqol", "resultschu", "resultsoutcomes",
    "resultseconomic", "serviceuse", "serviceusecost")
)
}
\arguments{
\item{model_data_ls}{Model data (a list), Default: NULL}

\item{platform_1L_chr}{Platform (a character vector of length one), Default: 'Intervention'}

\item{processed_ls}{Processed (a list), Default: NULL}

\item{regressions_ls}{Regressions (a list), Default: NULL}

\item{sim_results_ls}{Sim results (a list), Default: NULL}

\item{transformations_chr}{Transformations (a character vector), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("descriptives", "costadj", "costitem", "costsum", "costunit",
"minutes", "mnl-wait", "mnl-tx", "mnl-disc", "outcomes",
"outcomeslong", "paramscost", "paramsk10", "resultsaqol",
"resultschu", "resultsoutcomes", "resultseconomic", "serviceuse",
"serviceusecost")}
}
\value{
Data (an output object of multiple potential types)
}
\description{
make_report_data() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make report data. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
