% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_resources_tb}
\alias{make_resources_tb}
\title{Make resources tibble}
\usage{
make_resources_tb(
  resources_chr = character(0),
  categories_chr = character(0),
  derivation_chr = character(0),
  measures_chr = character(0),
  subcategories_chr = character(0),
  totals_chr = character(0)
)
}
\arguments{
\item{resources_chr}{Resources (a character vector), Default: character(0)}

\item{categories_chr}{Categories (a character vector), Default: character(0)}

\item{derivation_chr}{Derivation (a character vector), Default: character(0)}

\item{measures_chr}{Measures (a character vector), Default: character(0)}

\item{subcategories_chr}{Subcategories (a character vector), Default: character(0)}

\item{totals_chr}{Totals (a character vector), Default: character(0)}
}
\value{
Resources (a tibble)
}
\description{
make_resources_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make resources tibble. The function returns Resources (a tibble).
}
\keyword{internal}
