% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_results_summary}
\alias{make_results_summary}
\title{Make results summary}
\usage{
make_results_summary(
  X_Ready4useDyad,
  outcomes_chr,
  group_by_chr = character(0),
  min_cell_size_1L_int = 1L,
  threshold_1L_dbl = 96000,
  timestamp_1L_chr = get_timestamp(),
  utilities_chr = c("AQoL6D", "CHU9D")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{outcomes_chr}{Outcomes (a character vector)}

\item{group_by_chr}{Group by (a character vector), Default: character(0)}

\item{min_cell_size_1L_int}{Minimum cell size (an integer vector of length one), Default: 1}

\item{threshold_1L_dbl}{Threshold (a double vector of length one), Default: 96000}

\item{timestamp_1L_chr}{Timestamp (a character vector of length one), Default: get_timestamp()}

\item{utilities_chr}{Utilities (a character vector), Default: c("AQoL6D", "CHU9D")}
}
\value{
Results summary (a list)
}
\description{
make_results_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make results summary. The function returns Results summary (a list).
}
\keyword{internal}
