% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sim_env_ls}
\alias{make_sim_env_ls}
\title{Make sim environment list}
\usage{
make_sim_env_ls(
  sim_env_ls,
  append_ls = list(),
  discard_chr = "X_MimicConfiguration"
)
}
\arguments{
\item{sim_env_ls}{Sim (a list of environments)}

\item{append_ls}{Append (a list), Default: list()}

\item{discard_chr}{Discard (a character vector), Default: 'X_MimicConfiguration'}
}
\value{
Sim (a list of environments)
}
\description{
make_sim_env_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sim environment list. The function returns Sim (a list of environments).
}
\keyword{internal}
