% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_simulated_draws}
\alias{make_simulated_draws}
\title{Make simulated draws}
\usage{
make_simulated_draws(
  model_mdl,
  new_data_tb,
  sample_fn = rnorm,
  iterations_int = 1:100
)
}
\arguments{
\item{model_mdl}{Model (a model)}

\item{new_data_tb}{New data (a tibble)}

\item{sample_fn}{Sample (a function), Default: rnorm}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100}
}
\value{
Simulations (a data.frame)
}
\description{
make_simulated_draws() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make simulated draws. The function returns Simulations (a data.frame).
}
\keyword{internal}
