% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_simulation_fns_ls}
\alias{make_simulation_fns_ls}
\title{Make simulation functions list}
\usage{
make_simulation_fns_ls(
  type_1L_chr = c("all", "main", "processing", "sensitivity", "transformation"),
  comparator_fn = identity,
  extra_draws_fn = NULL,
  initialise_ls = make_initialise_ls(),
  intervention_fn = identity,
  sensitivities_ls = make_sensitivities_ls(),
  synthesis_fn = make_project_results_synthesis,
  transformation_ls = make_class_tfmns(),
  ...
)
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one), Default: c("all", "main", "processing", "sensitivity", "transformation")}

\item{comparator_fn}{Comparator (a function), Default: identity}

\item{extra_draws_fn}{Extra draws (a function), Default: NULL}

\item{initialise_ls}{Initialise (a list), Default: make_initialise_ls()}

\item{intervention_fn}{Intervention (a function), Default: identity}

\item{sensitivities_ls}{Sensitivities (a list), Default: make_sensitivities_ls()}

\item{synthesis_fn}{Synthesis (a function), Default: make_project_results_synthesis}

\item{transformation_ls}{Transformation (a list), Default: make_class_tfmns()}

\item{...}{Additional arguments}
}
\value{
Simulation functions (a list)
}
\description{
make_simulation_fns_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make simulation functions list. The function returns Simulation functions (a list).
}
\keyword{internal}
