% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_structural_vars}
\alias{make_structural_vars}
\title{Make structural variables}
\usage{
make_structural_vars()
}
\value{
Structural (a character vector)
}
\description{
make_structural_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make structural variables. The function returns Structural (a character vector).
}
\keyword{internal}
