% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_synthetic_tests}
\alias{make_synthetic_tests}
\title{Make synthetic tests}
\usage{
make_synthetic_tests(
  population_ls,
  model_data_ls = NULL,
  comparison_1L_chr = c("OutcomesJoinersImputed", "Joiners", "OutcomesJoiners",
    "Outcomes"),
  ...
)
}
\arguments{
\item{population_ls}{Population (a list)}

\item{model_data_ls}{Model data (a list), Default: NULL}

\item{comparison_1L_chr}{Comparison (a character vector of length one), Default: c("OutcomesJoinersImputed", "Joiners", "OutcomesJoiners", "Outcomes")}

\item{...}{Additional arguments}
}
\value{
Synthetic tests (a list)
}
\description{
make_synthetic_tests() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make synthetic tests. The function returns Synthetic tests (a list).
}
\keyword{internal}
