% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tfd_mds_outcomes_tb}
\alias{make_tfd_mds_outcomes_tb}
\title{Make transformed Minimum Dataset outcomes tibble}
\usage{
make_tfd_mds_outcomes_tb(
  processed_ls,
  outcomes_ls,
  jurisdiction_1L_chr = "Jurisdiction",
  missing_after_dtm = Sys.Date(),
  postcode_lup = NULL,
  referrals_1L_lgl = FALSE,
  reviews_1L_lgl = FALSE
)
}
\arguments{
\item{processed_ls}{Processed (a list)}

\item{outcomes_ls}{Outcomes (a list)}

\item{jurisdiction_1L_chr}{Jurisdiction (a character vector of length one), Default: 'Jurisdiction'}

\item{missing_after_dtm}{Missing after (a date vector), Default: Sys.Date()}

\item{postcode_lup}{Postcode (a lookup table), Default: NULL}

\item{referrals_1L_lgl}{Referrals (a logical vector of length one), Default: FALSE}

\item{reviews_1L_lgl}{Reviews (a logical vector of length one), Default: FALSE}
}
\value{
Data (a tibble)
}
\description{
make_tfd_mds_outcomes_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make transformed minimum dataset outcomes tibble. The function returns Data (a tibble).
}
\keyword{internal}
