% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_two_part_mdl}
\alias{make_two_part_mdl}
\title{Make two part model}
\usage{
make_two_part_mdl(
  data_tb,
  family_2_1L_chr = "Gamma(link = 'inverse')",
  link_1_1L_chr = "logit",
  x_part_1_chr,
  x_part_2_chr,
  y_1L_chr,
  ...
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{family_2_1L_chr}{Family 2 (a character vector of length one), Default: 'Gamma(link = 'inverse')'}

\item{link_1_1L_chr}{Link 1 (a character vector of length one), Default: 'logit'}

\item{x_part_1_chr}{X part 1 (a character vector)}

\item{x_part_2_chr}{X part 2 (a character vector)}

\item{y_1L_chr}{Y (a character vector of length one)}

\item{...}{Additional arguments}
}
\value{
Model (a model)
}
\description{
make_two_part_mdl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make two part model. The function returns Model (a model).
}
\keyword{internal}
