% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tx_mdl_confusion}
\alias{make_tx_mdl_confusion}
\title{Make treatment model confusion}
\usage{
make_tx_mdl_confusion(
  X_Ready4useDyad = ready4use::Ready4useDyad(),
  tx_mdls_ls,
  model_1L_int,
  treatment_vars_chr = c("treatment_status", "treatment_status_t2"),
  what_1L_chr = c("Waitlist", "Treatment", "Discharged")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{tx_mdls_ls}{Treatment models (a list)}

\item{model_1L_int}{Model (an integer vector of length one)}

\item{treatment_vars_chr}{Treatment variables (a character vector), Default: c("treatment_status", "treatment_status_t2")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("Waitlist", "Treatment", "Discharged")}
}
\value{
Confusion (a list)
}
\description{
make_tx_mdl_confusion() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make treatment model confusion. The function returns Confusion (a list).
}
