% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_unit_cost_params_tb}
\alias{make_unit_cost_params_tb}
\title{Make unit cost parameters tibble}
\usage{
make_unit_cost_params_tb(cost_per_mins_tb, new_ls = NULL)
}
\arguments{
\item{cost_per_mins_tb}{Cost per minimums (a tibble)}

\item{new_ls}{New (a list), Default: NULL}
}
\value{
Unit cost parameters (a tibble)
}
\description{
make_unit_cost_params_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make unit cost parameters tibble. The function returns Unit cost parameters (a tibble).
}
\keyword{internal}
