% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_utility_fns_ls}
\alias{make_utility_fns_ls}
\title{Make utility functions list}
\usage{
make_utility_fns_ls(
  add_to_ls = NULL,
  aqol8d_fn = add_aqol8d_from_k10,
  eq5d_fn = add_eq5d_from_draws,
  sf6d_fn = add_sf6d_from_draws,
  utilities_chr = c("AQoL8D", "EQ5D", "SF6D")
)
}
\arguments{
\item{add_to_ls}{Add to (a list), Default: NULL}

\item{aqol8d_fn}{Assessment of Quality of Life Eight Dimension (a function), Default: add_aqol8d_from_k10}

\item{eq5d_fn}{EQ5D (a function), Default: add_eq5d_from_draws}

\item{sf6d_fn}{Short Form - Six Dimension (a function), Default: add_sf6d_from_draws}

\item{utilities_chr}{Utilities (a character vector), Default: c("AQoL8D", "EQ5D", "SF6D")}
}
\value{
Utility functions (a list)
}
\description{
make_utility_fns_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make utility functions list. The function returns Utility functions (a list).
}
\keyword{internal}
