% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_utility_predictions_ds}
\alias{make_utility_predictions_ds}
\title{Make utility predictions dataset}
\usage{
make_utility_predictions_ds(
  X_Ready4useDyad = ready4use::Ready4useDyad(),
  Y_Ready4useDyad = ready4use::Ready4useDyad(),
  Z_Ready4useDyad = ready4use::Ready4useDyad(),
  model_mdl,
  utility_1L_chr = c("AQoL6D", "CHU9D"),
  follow_up_1L_int = 12,
  iterations_1L_int = 100L,
  join_with_chr = character(0),
  maintain_for_1L_int = 12,
  tfmn_1L_chr = "NTF",
  type_1L_chr = c("predict", "simulate"),
  with_1L_chr = "_sim_mean",
  what_1L_chr = c("old", "new")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{Z_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{model_mdl}{Model (a model)}

\item{utility_1L_chr}{Utility (a character vector of length one), Default: c("AQoL6D", "CHU9D")}

\item{follow_up_1L_int}{Follow up (an integer vector of length one), Default: 12}

\item{iterations_1L_int}{Iterations (an integer vector of length one), Default: 100}

\item{join_with_chr}{Join with (a character vector), Default: character(0)}

\item{maintain_for_1L_int}{Maintain for (an integer vector of length one), Default: 12}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("predict", "simulate")}

\item{with_1L_chr}{With (a character vector of length one), Default: '_sim_mean'}

\item{what_1L_chr}{What (a character vector of length one), Default: c("old", "new")}
}
\value{
Y (A dataset and data dictionary pair.)
}
\description{
make_utility_predictions_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make utility predictions dataset. The function is called for its side effects and does not return a value.
}
