% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_weeks_suffix}
\alias{make_weeks_suffix}
\title{Make weeks suffix}
\usage{
make_weeks_suffix(
  X_Ready4useDyad,
  adjustment_1L_dbl = 0,
  follow_up_1L_int = integer(0)
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{adjustment_1L_dbl}{Adjustment (a double vector of length one), Default: 0}

\item{follow_up_1L_int}{Follow up (an integer vector of length one), Default: integer(0)}
}
\value{
Suffix (a character vector of length one)
}
\description{
make_weeks_suffix() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make weeks suffix. The function returns Suffix (a character vector of length one).
}
\keyword{internal}
