% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_worker_types}
\alias{make_worker_types}
\title{Make worker types}
\usage{
make_worker_types(type_1L_chr = c("all", "medical"))
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one), Default: c("all", "medical")}
}
\value{
Worker types (a character vector)
}
\description{
make_worker_types() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make worker types. The function returns Worker types (a character vector).
}
