% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-MimicConfiguration}
\alias{manufacture-MimicConfiguration}
\alias{manufacture,MimicConfiguration-method}
\alias{manufacture-MimicRepos}
\alias{manufacture,MimicRepos-method}
\alias{manufacture-MimicInputs}
\alias{manufacture,MimicInputs-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{MimicConfiguration}(x, batch_1L_int = integer(0), what_1L_chr = c("draws_tb"))

\S4method{manufacture}{MimicRepos}(
  x,
  prefix_1L_chr = character(0),
  return_1L_chr = c("default", "batches", "files"),
  suffix_1L_chr = "",
  type_1L_chr = c("all", "batch_to", "draw_to"),
  what_1L_chr = c("sim_ws_dirs_chr"),
  ...
)

\S4method{manufacture}{MimicInputs}(x, what_1L_chr = c("inputs_ls"))
}
\arguments{
\item{x}{An object of class MimicInputs}

\item{batch_1L_int}{Batch (an integer vector of length one), Default: integer(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("inputs_ls")}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: character(0)}

\item{return_1L_chr}{Return (a character vector of length one), Default: c("default", "batches", "files")}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("all", "batch_to", "draw_to")}

\item{...}{Additional arguments}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
manufacture method applied to MimicConfiguration

manufacture method applied to MimicRepos

manufacture method applied to MimicInputs
}
