% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicActive.R
\docType{class}
\name{MimicActive-class}
\alias{MimicActive-class}
\alias{MimicActive}
\title{MimicActive}
\description{
Model active population (currently eligible for events).
}
\section{Slots}{

\describe{
\item{\code{x_Ready4useDyad}}{(an instance of the Ready4useDyad class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

