% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicAlgorithms.R
\docType{class}
\name{MimicAlgorithms-class}
\alias{MimicAlgorithms-class}
\alias{MimicAlgorithms}
\title{MimicAlgorithms}
\description{
The core set of functions that define simulation behaviour.
}
\section{Slots}{

\describe{
\item{\code{main_ls}}{Main (a list)}

\item{\code{processing_ls}}{Processing (a list)}

\item{\code{sensitivities_ls}}{Sensitivities (a list)}

\item{\code{transformations_ls}}{Transformations (a list)}

\item{\code{x_MimicUtility}}{(an instance of the MimicUtility class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

