% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicArguments.R
\docType{class}
\name{MimicArguments-class}
\alias{MimicArguments-class}
\alias{MimicArguments}
\title{MimicArguments}
\description{
Method argument data.
}
\section{Slots}{

\describe{
\item{\code{iterations_1L_lgl}}{Iterations (a logical vector of length one)}

\item{\code{derive_ls}}{Derive (a list)}

\item{\code{models_ls}}{Models (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

