% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicDerivations.R
\docType{class}
\name{MimicDerivations-class}
\alias{MimicDerivations-class}
\alias{MimicDerivations}
\title{MimicDerivations}
\description{
Data for deriving method arguments.
}
\section{Slots}{

\describe{
\item{\code{method_1L_chr}}{Method (a character vector of length one)}

\item{\code{args_env_ls}}{Arguments (a list of environments)}

\item{\code{args_fixed_ls}}{Arguments fixed (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

