% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicEligible.R
\docType{class}
\name{MimicEligible-class}
\alias{MimicEligible-class}
\alias{MimicEligible}
\title{MimicEligible}
\description{
Model event eligibility logic data.
}
\section{Slots}{

\describe{
\item{\code{ineligible_1L_chr}}{Ineligible (a character vector of length one)}

\item{\code{functions_ls}}{Functions (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

