% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicEvent.R
\docType{class}
\name{MimicEvent-class}
\alias{MimicEvent-class}
\alias{MimicEvent}
\title{MimicEvent}
\description{
Model event scheduling and event logic data.
}
\section{Slots}{

\describe{
\item{\code{x_MimicSchedule}}{(an instance of the MimicSchedule class)}

\item{\code{y_MimicTrigger}}{(an instance of the MimicTrigger class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

