% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicEventsList.R
\docType{class}
\name{MimicEventsList-class}
\alias{MimicEventsList-class}
\alias{MimicEventsList}
\title{MimicEventsList}
\description{
Model event scheduling and event logic data.
}
\section{Slots}{

\describe{
\item{\code{events_ls}}{(a list)}

\item{\code{last_1L_chr}}{(a character vector of length one)}

\item{\code{main_chr}}{(a character vector )}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

