% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicInputs.R
\docType{class}
\name{MimicInputs-class}
\alias{MimicInputs-class}
\alias{MimicInputs}
\title{MimicInputs}
\description{
Model input data (regression models, population and parameters).
}
\section{Slots}{

\describe{
\item{\code{models_ls}}{Models (a list)}

\item{\code{x_Ready4useDyad}}{(an instance of the Ready4useDyad class)}

\item{\code{y_Ready4useDyad}}{(an instance of the Ready4useDyad class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

