% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicPopulation.R
\docType{class}
\name{MimicPopulation-class}
\alias{MimicPopulation-class}
\alias{MimicPopulation}
\title{MimicPopulation}
\description{
Model dynamic population data container (specifying population subgroups who are active and eligible for the next model event,  who are temporariliy inactive but are still in the model or who have exited the model).
}
\section{Slots}{

\describe{
\item{\code{x_MimicActive}}{(an instance of the MimicActive class)}

\item{\code{y_Ready4useDyad}}{(an instance of the Ready4useDyad class)}

\item{\code{z_Ready4useDyad}}{(an instance of the Ready4useDyad class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

