% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicRepos.R
\docType{class}
\name{MimicRepos-class}
\alias{MimicRepos-class}
\alias{MimicRepos}
\title{MimicRepos}
\description{
Local and remote repositories for model input and output data.
}
\section{Slots}{

\describe{
\item{\code{batch_to_1L_chr}}{Batch to (a character vector of length one)}

\item{\code{divider_1L_chr}}{Divider (a character vector of length one)}

\item{\code{outp_data_dir_1L_chr}}{Output data directory (a character vector of length one)}

\item{\code{path_to_keys_1L_chr}}{Path to keys (a character vector of length one)}

\item{\code{path_to_output_1L_chr}}{Path to output (a character vector of length one)}

\item{\code{path_to_param_data_1L_chr}}{Path to parameter data (a character vector of length one)}

\item{\code{path_to_private_1L_chr}}{Path to private (a character vector of length one)}

\item{\code{processed_dir_1L_chr}}{Processed directory (a character vector of length one)}

\item{\code{r_dir_1L_chr}}{R directory (a character vector of length one)}

\item{\code{raw_dir_1L_chr}}{Raw directory (a character vector of length one)}

\item{\code{reports_dir_1L_chr}}{Reports directory (a character vector of length one)}

\item{\code{x_Ready4useRepos}}{(an instance of the Ready4useRepos class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

