% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicSchedule.R
\docType{class}
\name{MimicSchedule-class}
\alias{MimicSchedule-class}
\alias{MimicSchedule}
\title{MimicSchedule}
\description{
Model event scheduling data.
}
\section{Slots}{

\describe{
\item{\code{assert_1L_lgl}}{Assert (a logical vector of length one)}

\item{\code{event_1L_chr}}{Event (a character vector of length one)}

\item{\code{functions_ls}}{Functions (a list)}

\item{\code{step_dtm}}{Step (a date vector)}

\item{\code{update_type_1L_chr}}{Update type (a character vector of length one)}

\item{\code{use_1L_chr}}{Use (a character vector of length one)}

\item{\code{validate_chr}}{Validate (a character vector)}

\item{\code{x_MimicArguments}}{(an instance of the MimicArguments class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

