% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicTrigger.R
\docType{class}
\name{MimicTrigger-class}
\alias{MimicTrigger-class}
\alias{MimicTrigger}
\title{MimicTrigger}
\description{
Model event logic data.
}
\section{Slots}{

\describe{
\item{\code{assert_1L_lgl}}{Assert (a logical vector of length one)}

\item{\code{event_1L_chr}}{Event (a character vector of length one)}

\item{\code{functions_ls}}{Functions (a list)}

\item{\code{use_1L_chr}}{Use (a character vector of length one)}

\item{\code{validate_chr}}{Validate (a character vector)}

\item{\code{x_MimicArguments}}{(an instance of the MimicArguments class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

