% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicUtility.R
\docType{class}
\name{MimicUtility-class}
\alias{MimicUtility-class}
\alias{MimicUtility}
\title{MimicUtility}
\description{
Utility algorithm data (instrument names, modules, mapping functions and transformation functions).
}
\section{Slots}{

\describe{
\item{\code{names_chr}}{Names (a character vector)}

\item{\code{data_ls}}{Data (a list)}

\item{\code{mapping_ls}}{Mapping (a list)}

\item{\code{transformations_ls}}{Transformations (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

