% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicVariables.R
\docType{class}
\name{MimicVariables-class}
\alias{MimicVariables-class}
\alias{MimicVariables}
\title{MimicVariables}
\description{
Model outcome and resource use variables metadata.
}
\section{Slots}{

\describe{
\item{\code{outcomes_tb}}{Outcomes (a tibble)}

\item{\code{resources_tb}}{Resources (a tibble)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

