% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_economic_results}
\alias{plot_economic_results}
\title{Plot economic results}
\usage{
plot_economic_results(
  economic_results_ls,
  what_1L_chr,
  alpha_1L_dbl = 0.8,
  colour_1L_chr = ready4use::get_colour_codes(2)[2],
  size_1L_dbl = 3,
  threshold_1L_dbl = 96000,
  title_1L_chr = " ",
  type_1L_chr = c("cep", "ceac", "evi"),
  x_limits_dbl = numeric(0),
  ...
)
}
\arguments{
\item{economic_results_ls}{Economic results (a list)}

\item{what_1L_chr}{What (a character vector of length one)}

\item{alpha_1L_dbl}{Alpha (a double vector of length one), Default: 0.8}

\item{colour_1L_chr}{Colour (a character vector of length one), Default: ready4use::get_colour_codes(2)\link{2}}

\item{size_1L_dbl}{Size (a double vector of length one), Default: 3}

\item{threshold_1L_dbl}{Threshold (a double vector of length one), Default: 96000}

\item{title_1L_chr}{Title (a character vector of length one), Default: ' '}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("cep", "ceac", "evi")}

\item{x_limits_dbl}{X limits (a double vector), Default: numeric(0)}

\item{...}{Additional arguments}
}
\value{
Plot (a plot)
}
\description{
plot_economic_results() is a Plot function that plots data. Specifically, this function implements an algorithm to plot economic results. The function returns Plot (a plot).
}
\keyword{internal}
