% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_pooled}
\alias{plot_pooled}
\title{Plot pooled}
\usage{
plot_pooled(
  pooled_fits_ls,
  what_1L_chr,
  distributions_chr = "best",
  pool_1L_lgl = TRUE,
  ...
)
}
\arguments{
\item{pooled_fits_ls}{Pooled fits (a list)}

\item{what_1L_chr}{What (a character vector of length one)}

\item{distributions_chr}{Distributions (a character vector), Default: 'best'}

\item{pool_1L_lgl}{Pool (a logical vector of length one), Default: TRUE}

\item{...}{Additional arguments}
}
\value{
Plot (a plot)
}
\description{
plot_pooled() is a Plot function that plots data. Specifically, this function implements an algorithm to plot pooled. The function returns Plot (a plot).
}
\keyword{internal}
