% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_digital_pathway}
\alias{predict_digital_pathway}
\title{Predict digital pathway}
\usage{
predict_digital_pathway(
  inputs_ls,
  base_for_rates_int = c(1000L, 1L, 1L),
  draws_tb = NULL,
  iterations_int = 1:100L,
  horizon_dtm = lubridate::years(1),
  modifiable_chr = c("treatment_status", "Minutes", "k10", "AQoL6D", "CHU9D"),
  tfmn_ls = make_class_tfmns(),
  tx_duration_dtm = lubridate::weeks(12),
  seed_1L_int = 2001L,
  start_dtm = Sys.Date()
)
}
\arguments{
\item{inputs_ls}{Inputs (a list)}

\item{base_for_rates_int}{Base for rates (an integer vector), Default: c(1000L, 1L, 1L)}

\item{draws_tb}{Draws (a tibble), Default: NULL}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100L}

\item{horizon_dtm}{Horizon (a date vector), Default: lubridate::years(1)}

\item{modifiable_chr}{Modifiable (a character vector), Default: c("treatment_status", "Minutes", "k10", "AQoL6D", "CHU9D")}

\item{tfmn_ls}{Transformation (a list), Default: make_class_tfmns()}

\item{tx_duration_dtm}{Treatment duration (a date vector), Default: lubridate::weeks(12)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 2001}

\item{start_dtm}{Start (a date vector), Default: Sys.Date()}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
predict_digital_pathway() is a Predict function that applies a model to make predictions. Specifically, this function implements an algorithm to predict digital pathway. The function is called for its side effects and does not return a value.
}
\keyword{internal}
