% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_from_pool}
\alias{predict_from_pool}
\title{Predict from pool}
\usage{
predict_from_pool(
  pooled_xx,
  as_1L_chr = c("vector", "histogram", "summary"),
  adjustment_1L_dbl = numeric(0),
  distributions_chr = "best",
  maximum_1L_dbl = Inf,
  minimum_1L_dbl = -Inf,
  n_1L_int = 100L,
  quantiles_dbl = c(0.05, 0.25, 0.5, 0.75, 0.95),
  resample_1L_lgl = F,
  seed_1L_int = 2001L,
  title_1L_chr = "Pooled prediction distribution",
  type_1L_chr = c("r", "q", "p"),
  x_label_1L_chr = "Predictions",
  values_dbl = numeric(0),
  weights_dbl = 1,
  what_1L_chr = character(0),
  ...
)
}
\arguments{
\item{pooled_xx}{Pooled (an output object of multiple potential types)}

\item{as_1L_chr}{As (a character vector of length one), Default: c("vector", "histogram", "summary")}

\item{adjustment_1L_dbl}{Adjustment (a double vector of length one), Default: numeric(0)}

\item{distributions_chr}{Distributions (a character vector), Default: 'best'}

\item{maximum_1L_dbl}{Maximum (a double vector of length one), Default: Inf}

\item{minimum_1L_dbl}{Minimum (a double vector of length one), Default: -Inf}

\item{n_1L_int}{N (an integer vector of length one), Default: 100}

\item{quantiles_dbl}{Quantiles (a double vector), Default: c(0.05, 0.25, 0.5, 0.75, 0.95)}

\item{resample_1L_lgl}{Resample (a logical vector of length one), Default: F}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 2001}

\item{title_1L_chr}{Title (a character vector of length one), Default: 'Pooled prediction distribution'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("r", "q", "p")}

\item{x_label_1L_chr}{X label (a character vector of length one), Default: 'Predictions'}

\item{values_dbl}{Values (a double vector), Default: numeric(0)}

\item{weights_dbl}{Weights (a double vector), Default: 1}

\item{what_1L_chr}{What (a character vector of length one), Default: character(0)}

\item{...}{Additional arguments}
}
\value{
Predictions (an output object of multiple potential types)
}
\description{
predict_from_pool() is a Predict function that applies a model to make predictions. Specifically, this function implements an algorithm to predict from pool. The function returns Predictions (an output object of multiple potential types).
}
