% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_with_sim}
\alias{predict_with_sim}
\title{Predict with sim}
\usage{
predict_with_sim(
  inputs_ls,
  base_for_rates_int = c(1000L, 1L, 1L),
  iterations_ls = list(1:100L),
  horizon_dtm = lubridate::years(1),
  modifiable_chr = c("treatment_status", "Minutes", "k10", "AQoL6D", "CHU9D"),
  purge_1L_lgl = TRUE,
  scale_1L_int = 10L,
  seed_1L_int = 2001L,
  sensitivities_ls = make_sensitivities_ls(),
  start_dtm = Sys.Date(),
  tfmn_ls = make_class_tfmns(),
  tx_duration_dtm = lubridate::weeks(12),
  type_1L_chr = c("D", "AB", "C"),
  utilities_chr = c("AQoL6D", "CHU9D"),
  write_to_1L_chr = character(0)
)
}
\arguments{
\item{inputs_ls}{Inputs (a list)}

\item{base_for_rates_int}{Base for rates (an integer vector), Default: c(1000L, 1L, 1L)}

\item{iterations_ls}{Iterations (a list), Default: list(1:100L)}

\item{horizon_dtm}{Horizon (a date vector), Default: lubridate::years(1)}

\item{modifiable_chr}{Modifiable (a character vector), Default: c("treatment_status", "Minutes", "k10", "AQoL6D", "CHU9D")}

\item{purge_1L_lgl}{Purge (a logical vector of length one), Default: TRUE}

\item{scale_1L_int}{Scale (an integer vector of length one), Default: 10}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 2001}

\item{sensitivities_ls}{Sensitivities (a list), Default: make_sensitivities_ls()}

\item{start_dtm}{Start (a date vector), Default: Sys.Date()}

\item{tfmn_ls}{Transformation (a list), Default: make_class_tfmns()}

\item{tx_duration_dtm}{Treatment duration (a date vector), Default: lubridate::weeks(12)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("D", "AB", "C")}

\item{utilities_chr}{Utilities (a character vector), Default: c("AQoL6D", "CHU9D")}

\item{write_to_1L_chr}{Write to (a character vector of length one), Default: character(0)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
predict_with_sim() is a Predict function that applies a model to make predictions. Specifically, this function implements an algorithm to predict with sim. The function is called for its side effects and does not return a value.
}
