% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_cmprsns}
\alias{print_cmprsns}
\title{Print comparisons}
\usage{
print_cmprsns(
  data_xx,
  output_type_1L_chr = c("HTML", "PDF", "Word"),
  what_1L_chr = c("gtsummary", "df", "null")
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: c("HTML", "PDF", "Word")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("gtsummary", "df", "null")}
}
\value{
Table (an output object of multiple potential types)
}
\description{
print_cmprsns() is a Print function that prints output to console. Specifically, this function implements an algorithm to print comparisons. The function returns Table (an output object of multiple potential types).
}
\keyword{internal}
