% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_errors}
\alias{print_errors}
\title{Print errors}
\usage{
print_errors(
  X_Ready4useDyad,
  vars_chr,
  assert_1L_lgl = FALSE,
  invalid_fn = function(x) (is.na(x) | is.nan(x) | is.null(x) | x == -Inf | x == Inf)
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{vars_chr}{Variables (a character vector)}

\item{assert_1L_lgl}{Assert (a logical vector of length one), Default: FALSE}

\item{invalid_fn}{Invalid (a function), Default: function(x) (is.na(x) | is.nan(x) | is.null(x) | x == -Inf |
x == Inf)}
}
\value{
No return value, called for side effects.
}
\description{
print_errors() is a Print function that prints output to console. Specifically, this function implements an algorithm to print errors. The function is called for its side effects and does not return a value.
}
