% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure-MimicConfiguration}
\alias{procure-MimicConfiguration}
\alias{procure,MimicConfiguration-method}
\title{Procure data from a model module}
\usage{
\S4method{procure}{MimicConfiguration}(
  x,
  empty_xx = NULL,
  match_value_xx = NULL,
  target_1L_chr = character(0),
  type_1L_chr = "Arm",
  what_1L_chr = c("arm", "event")
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{empty_xx}{Empty (an output object of multiple potential types), Default: NULL}

\item{match_value_xx}{Match value (an output object of multiple potential types), Default: NULL}

\item{target_1L_chr}{Target (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'Arm'}

\item{what_1L_chr}{What (a character vector of length one), Default: c("arm")}
}
\value{
Object (an output object of multiple potential types)
}
\description{
procure method applied to MimicConfiguration
}
