% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ratify.R, R/mthd_renew.R
\name{ratify-MimicRepos}
\alias{ratify-MimicRepos}
\alias{ratify,MimicRepos-method}
\alias{renew-MimicActive}
\alias{renew,MimicActive-method}
\alias{renew-MimicPopulation}
\alias{renew,MimicPopulation-method}
\alias{renew-MimicConfiguration}
\alias{renew,MimicConfiguration-method}
\title{Ratify that input or output data meet validity criteria}
\usage{
\S4method{ratify}{MimicRepos}(
  x,
  batches_int = integer(0),
  gh_token_1L_chr = "",
  key_1L_chr = NULL,
  match_xx = NULL,
  prefix_1L_chr = character(0),
  suffix_1L_chr = "",
  remote_fls_chr = NA_character_,
  type_1L_chr = c("MimicInputs", "ParamDraws", "list", "element"),
  what_chr = character(0),
  ...
)

\S4method{renew}{MimicActive}(
  x,
  batch_1L_int = integer(0),
  env_ls = list(),
  event_1L_chr = character(0),
  type_1L_chr = c("trigger", "customise", "filter", "schedule"),
  X_MimicConfiguration = MimicConfiguration(),
  ...
)

\S4method{renew}{MimicPopulation}(
  x,
  batch_1L_int = integer(0),
  env_ls = list(),
  event_1L_chr = character(0),
  population_ls = NULL,
  type_1L_chr = c("trigger", "customise", "filter", "event", "reset", "schedule",
    "switch", "transform"),
  use_1L_chr = "Y",
  what_1L_chr = character(0),
  X_MimicConfiguration = MimicConfiguration(),
  ...
)

\S4method{renew}{MimicConfiguration}(x, env_ls = list(), what_1L_chr = c("legacy"), ...)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{batches_int}{Batches (an integer vector), Default: integer(0)}

\item{gh_token_1L_chr}{GitHub token (a character vector of length one), Default: ""}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{match_xx}{Match (an object), Default: NULL}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: character(0)}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ""}

\item{remote_fls_chr}{Remote files (a character vector), Default: NA_character_}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("trigger", "customise", "filter", "event", "reset", "schedule",
"switch", "transform")}

\item{what_chr}{What (a character vector), Default:  character(0)}

\item{...}{Additional arguments}

\item{batch_1L_int}{Batch (an integer vector of length one), Default: integer(0)}

\item{env_ls}{Environment list (a list of environments), Default: list()}

\item{event_1L_chr}{Event (a character vector of length one), Default: character(0)}

\item{X_MimicConfiguration}{PARAM_DESCRIPTION, Default: MimicConfiguration()}

\item{population_ls}{Population (a list), Default: NULL}

\item{use_1L_chr}{Use (a character vector of length one), Default: 'Y'}

\item{what_1L_chr}{What (a character vector of length one), Default: c("legacy")}
}
\value{
x (An object of class MimicRepos)

x (An object of class MimicActive)

x (An object of class MimicPopulation)

x (An object of class MimicConfiguration)
}
\description{
ratify method applied to MimicRepos

renew method applied to MimicActive

renew method applied to MimicPopulation

renew method applied to MimicConfiguration
}
