% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_to_wide}
\alias{transform_ds_to_wide}
\title{Transform dataset to wide}
\usage{
transform_ds_to_wide(
  X_Ready4useDyad,
  processed_ls,
  join_before_dtm = NULL,
  key_vars_chr = make_project_keys(),
  max_periods_1L_int = integer(0),
  max_tenure_1L_dbl = numeric(0),
  metric_var_1L_chr = "Minutes",
  patterns_ls = list(c("[[:space:]]", ""), c("NA", "MISSING")),
  period_var_1L_chr = "Year",
  separation_after_dbl = 3,
  service_var_1L_chr = "primary_purpose"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{processed_ls}{Processed (a list)}

\item{join_before_dtm}{Join before (a date vector), Default: NULL}

\item{key_vars_chr}{Key variables (a character vector), Default: make_project_keys()}

\item{max_periods_1L_int}{Maximum periods (an integer vector of length one), Default: integer(0)}

\item{max_tenure_1L_dbl}{Maximum tenure (a double vector of length one), Default: numeric(0)}

\item{metric_var_1L_chr}{Metric variable (a character vector of length one), Default: 'Minutes'}

\item{patterns_ls}{Patterns (a list), Default: list(c("[\link{:space:}]", ""), c("NA", "MISSING"))}

\item{period_var_1L_chr}{Period variable (a character vector of length one), Default: 'Year'}

\item{separation_after_dbl}{Separation after (a double vector), Default: 3}

\item{service_var_1L_chr}{Service variable (a character vector of length one), Default: 'primary_purpose'}
}
\value{
Y (A dataset and data dictionary pair.)
}
\description{
transform_ds_to_wide() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset to wide. The function is called for its side effects and does not return a value.
}
\keyword{internal}
