% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{transform_integer_dates}
\alias{transform_integer_dates}
\title{Transform integer dates}
\usage{
transform_integer_dates(dates_int)
}
\arguments{
\item{dates_int}{Dates (an integer vector)}
}
\value{
Dates (a date vector)
}
\description{
transform_integer_dates() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform integer dates. The function returns Dates (a date vector).
}
\keyword{internal}
