% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_to_long_results}
\alias{transform_to_long_results}
\title{Transform to long results}
\usage{
transform_to_long_results(
  X_Ready4useDyad,
  var_1L_chr,
  add_means_1L_lgl = TRUE,
  tidy_1L_lgl = TRUE
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{var_1L_chr}{Variable (a character vector of length one)}

\item{add_means_1L_lgl}{Add means (a logical vector of length one), Default: TRUE}

\item{tidy_1L_lgl}{Tidy (a logical vector of length one), Default: TRUE}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
transform_to_long_results() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to long results. The function is called for its side effects and does not return a value.
}
\keyword{internal}
