% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_to_min_and_max}
\alias{transform_to_min_and_max}
\title{Transform to minimum and maximum}
\usage{
transform_to_min_and_max(
  X_Ready4useDyad,
  vars_chr,
  max_1L_dbl = 0.999999,
  min_1L_dbl = 1e-06
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{vars_chr}{Variables (a character vector)}

\item{max_1L_dbl}{Maximum (a double vector of length one), Default: 0.999999}

\item{min_1L_dbl}{Minimum (a double vector of length one), Default: 1e-06}
}
\value{
Y (A dataset and data dictionary pair.)
}
\description{
transform_to_min_and_max() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to minimum and maximum. The function is called for its side effects and does not return a value.
}
\keyword{internal}
