% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_to_remove_duplicates}
\alias{transform_to_remove_duplicates}
\title{Transform to remove duplicates}
\usage{
transform_to_remove_duplicates(
  data_tb,
  group_by_chr = c("episode_key", "collection_occasion_date", "reason_for_collection"),
  uid_1L_chr = "episode_key"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{group_by_chr}{Group by (a character vector), Default: c("episode_key", "collection_occasion_date", "reason_for_collection")}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'episode_key'}
}
\value{
Data (a tibble)
}
\description{
transform_to_remove_duplicates() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to remove duplicates. The function returns Data (a tibble).
}
\keyword{internal}
