% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_tx_factor}
\alias{transform_tx_factor}
\title{Transform treatment factor}
\usage{
transform_tx_factor(
  data_tb,
  treatment_vars_chr = c("treatment_status", "treatment_status_t2"),
  what_1L_chr = c("Waitlist", "Treatment", "Discharged")
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{treatment_vars_chr}{Treatment variables (a character vector), Default: c("treatment_status", "treatment_status_t2")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("Waitlist", "Treatment", "Discharged")}
}
\value{
Data (a tibble)
}
\description{
transform_tx_factor() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform treatment factor. The function returns Data (a tibble).
}
\keyword{internal}
