% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_arguments_ls}
\alias{update_arguments_ls}
\title{Update arguments list}
\usage{
update_arguments_ls(args_ls, function_fn)
}
\arguments{
\item{args_ls}{Arguments (a list)}

\item{function_fn}{Function (a function)}
}
\value{
Arguments (a list)
}
\description{
update_arguments_ls() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update arguments list. The function returns Arguments (a list).
}
\keyword{internal}
