% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_current_date}
\alias{update_current_date}
\title{Update current date}
\usage{
update_current_date(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_current_date() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update current date. The function is called for its side effects and does not return a value.
}
\keyword{internal}
