% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_episodes_lup}
\alias{update_episodes_lup}
\title{Update episodes lookup table}
\usage{
update_episodes_lup(
  episodes_lup_tb,
  dates_chr,
  part_one_1L_chr,
  program_true_1L_chr
)
}
\arguments{
\item{episodes_lup_tb}{Episodes lookup table (a tibble)}

\item{dates_chr}{Dates (a character vector)}

\item{part_one_1L_chr}{Part one (a character vector of length one)}

\item{program_true_1L_chr}{Program true (a character vector of length one)}
}
\value{
Episodes lookup table (a tibble)
}
\description{
update_episodes_lup() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update episodes lookup table. The function returns Episodes lookup table (a tibble).
}
\keyword{internal}
