% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_gender}
\alias{update_gender}
\title{Update gender}
\usage{
update_gender(data_tb)
}
\arguments{
\item{data_tb}{Data (a tibble)}
}
\value{
Data (a tibble)
}
\description{
update_gender() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update gender. The function returns Data (a tibble).
}
\keyword{internal}
