% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_intervention_name}
\alias{update_intervention_name}
\title{Update intervention name}
\usage{
update_intervention_name(
  data_tb,
  new_1L_chr = "Comparator",
  old_1L_chr = "FlexPsych",
  var_nm_1L_chr = "Intervention"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{new_1L_chr}{New (a character vector of length one), Default: 'Comparator'}

\item{old_1L_chr}{Old (a character vector of length one), Default: 'FlexPsych'}

\item{var_nm_1L_chr}{Variable name (a character vector of length one), Default: 'Intervention'}
}
\value{
Data (a tibble)
}
\description{
update_intervention_name() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update intervention name. The function returns Data (a tibble).
}
