% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_k10_event_schedule}
\alias{update_k10_event_schedule}
\title{Update K10 event schedule}
\usage{
update_k10_event_schedule(X_Ready4useDyad, type_1L_chr = c("Model", "Table"))
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("Model", "Table")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_k10_event_schedule() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update k10 event schedule. The function is called for its side effects and does not return a value.
}
\keyword{internal}
