% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_mds_modelling_ds}
\alias{update_mds_modelling_ds}
\title{Update Minimum Dataset modelling dataset}
\usage{
update_mds_modelling_ds(
  X_Ready4useDyad,
  imputations_int = 1,
  sample_ls = NULL,
  filter_true_1L_chr = "FlexPsych"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{imputations_int}{Imputations (an integer vector), Default: 1}

\item{sample_ls}{Sample (a list), Default: NULL}

\item{filter_true_1L_chr}{Filter true (a character vector of length one), Default: 'FlexPsych'}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_mds_modelling_ds() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update minimum dataset modelling dataset. The function is called for its side effects and does not return a value.
}
\keyword{internal}
