% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_minute_var_nms}
\alias{update_minute_var_nms}
\title{Update minute variable names}
\usage{
update_minute_var_nms(data_tb, type_1L_chr = c("undo", "do"))
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("undo", "do")}
}
\value{
Data (a tibble)
}
\description{
update_minute_var_nms() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update minute variable names. The function returns Data (a tibble).
}
\keyword{internal}
