% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_mismatched_vars}
\alias{update_mismatched_vars}
\title{Update mismatched variables}
\usage{
update_mismatched_vars(
  model_dyad_ls = make_model_dyad_ls,
  type_1L_chr = c("drop", "rename")
)
}
\arguments{
\item{model_dyad_ls}{Model dyad (a list), Default: make_model_dyad_ls}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("drop", "rename")}
}
\value{
Model dyad (a list)
}
\description{
update_mismatched_vars() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update mismatched variables. The function returns Model dyad (a list).
}
\keyword{internal}
