% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_order}
\alias{update_order}
\title{Update order}
\usage{
update_order(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_order() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update order. The function is called for its side effects and does not return a value.
}
\keyword{internal}
