% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_population_classes}
\alias{update_population_classes}
\title{Update population classes}
\usage{
update_population_classes(X_Ready4useDyad, tfmn_ls = NULL)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{tfmn_ls}{Transformation (a list), Default: NULL}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_population_classes() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update population classes. The function is called for its side effects and does not return a value.
}
