% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_predictions_ds}
\alias{update_predictions_ds}
\title{Update predictions dataset}
\usage{
update_predictions_ds(
  Y_Ready4useDyad,
  adjustment_1L_dbl = 0,
  var_1L_chr = character(0),
  do_int = 1:5,
  follow_up_1L_int = 12L,
  maintain_for_1L_int = 0L,
  suffixes_chr = c("_YR1", "_YR1_S1", "_YR1_S2"),
  tfmn_1L_chr = "NTF",
  utility_1L_chr = c("AQoL6D", "CHU9D"),
  with_1L_chr = "_sim_mean"
)
}
\arguments{
\item{Y_Ready4useDyad}{PARAM_DESCRIPTION}

\item{adjustment_1L_dbl}{Adjustment (a double vector of length one), Default: 0}

\item{var_1L_chr}{Variable (a character vector of length one), Default: character(0)}

\item{do_int}{Do (an integer vector), Default: 1:5}

\item{follow_up_1L_int}{Follow up (an integer vector of length one), Default: 12}

\item{maintain_for_1L_int}{Maintain for (an integer vector of length one), Default: 0}

\item{suffixes_chr}{Suffixes (a character vector), Default: c("_YR1", "_YR1_S1", "_YR1_S2")}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{utility_1L_chr}{Utility (a character vector of length one), Default: c("AQoL6D", "CHU9D")}

\item{with_1L_chr}{With (a character vector of length one), Default: '_sim_mean'}
}
\value{
Y (A dataset and data dictionary pair.)
}
\description{
update_predictions_ds() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update predictions dataset. The function is called for its side effects and does not return a value.
}
\keyword{internal}
