% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_previous}
\alias{update_previous}
\title{Update previous}
\usage{
update_previous(
  X_Ready4useDyad,
  modifiable_chr = character(0),
  pattern_1L_chr = "{col}_previous"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{modifiable_chr}{Modifiable (a character vector), Default: character(0)}

\item{pattern_1L_chr}{Pattern (a character vector of length one), Default: '{col}_previous'}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_previous() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update previous. The function is called for its side effects and does not return a value.
}
\keyword{internal}
