% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_processed_tb}
\alias{update_processed_tb}
\title{Update processed tibble}
\usage{
update_processed_tb(
  data_tb,
  first_eight_1L_lgl = NA,
  program_1L_chr = NA_character_
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{first_eight_1L_lgl}{First eight (a logical vector of length one), Default: NA}

\item{program_1L_chr}{Program (a character vector of length one), Default: 'NA'}
}
\value{
Data (a tibble)
}
\description{
update_processed_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update processed tibble. The function returns Data (a tibble).
}
\keyword{internal}
