% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_project_2_param_names}
\alias{update_project_2_param_names}
\title{Update project 2 parameter names}
\usage{
update_project_2_param_names(params_tb)
}
\arguments{
\item{params_tb}{Parameters (a tibble)}
}
\value{
Parameters (a tibble)
}
\description{
update_project_2_param_names() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update project 2 parameter names. The function returns Parameters (a tibble).
}
\keyword{internal}
