% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_project_test_cmprsns}
\alias{update_project_test_cmprsns}
\title{Update project test comparisons}
\usage{
update_project_test_cmprsns(X_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_project_test_cmprsns() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update project test comparisons. The function is called for its side effects and does not return a value.
}
\keyword{internal}
