% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_providers_tb}
\alias{update_providers_tb}
\title{Update providers tibble}
\usage{
update_providers_tb(providers_tb, var_1L_chr = "practitioner_category")
}
\arguments{
\item{providers_tb}{Providers (a tibble)}

\item{var_1L_chr}{Variable (a character vector of length one), Default: 'practitioner_category'}
}
\value{
Providers (a tibble)
}
\description{
update_providers_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update providers tibble. The function returns Providers (a tibble).
}
