% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_qalys}
\alias{update_qalys}
\title{Update Quality Adjusted Life Years}
\usage{
update_qalys(
  X_Ready4useDyad,
  add_sensitivity_1L_lgl = FALSE,
  adjustment_1L_dbl = 0,
  follow_up_1L_int = integer(0),
  maintain_for_1L_int = 0,
  tidy_1L_lgl = FALSE,
  utilities_chr = c("AQoL6D", "CHU9D")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{add_sensitivity_1L_lgl}{Add sensitivity (a logical vector of length one), Default: FALSE}

\item{adjustment_1L_dbl}{Adjustment (a double vector of length one), Default: 0}

\item{follow_up_1L_int}{Follow up (an integer vector of length one), Default: integer(0)}

\item{maintain_for_1L_int}{Maintain for (an integer vector of length one), Default: 0}

\item{tidy_1L_lgl}{Tidy (a logical vector of length one), Default: FALSE}

\item{utilities_chr}{Utilities (a character vector), Default: c("AQoL6D", "CHU9D")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_qalys() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update quality adjusted life years. The function is called for its side effects and does not return a value.
}
\keyword{internal}
