% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_scenario_names}
\alias{update_scenario_names}
\title{Update scenario names}
\usage{
update_scenario_names(
  forecasts_tb,
  after_1L_chr = character(0),
  before_1L_chr = character(0),
  prefix_1L_chr = "scenario_",
  reference_1L_chr = "Status quo",
  tfmn_1_fn = as.numeric,
  tfmn_2_fn = scales::percent
)
}
\arguments{
\item{forecasts_tb}{Forecasts (a tibble)}

\item{after_1L_chr}{After (a character vector of length one), Default: character(0)}

\item{before_1L_chr}{Before (a character vector of length one), Default: character(0)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'scenario_'}

\item{reference_1L_chr}{Reference (a character vector of length one), Default: 'Status quo'}

\item{tfmn_1_fn}{Transformation 1 (a function), Default: as.numeric}

\item{tfmn_2_fn}{Transformation 2 (a function), Default: scales::percent}
}
\value{
Forecasts (a tibble)
}
\description{
update_scenario_names() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update scenario names. The function returns Forecasts (a tibble).
}
\keyword{internal}
