% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_scheduled_date}
\alias{update_scheduled_date}
\title{Update scheduled date}
\usage{
update_scheduled_date(X_Ready4useDyad, type_1L_chr = c("End"))
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("End")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_scheduled_date() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update scheduled date. The function is called for its side effects and does not return a value.
}
\keyword{internal}
