% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_test_ds}
\alias{update_test_ds}
\title{Update test dataset}
\usage{
update_test_ds(
  X_Ready4useDyad,
  modifiable_chr,
  pattern_1L_chr = "{col}_1_year",
  period_dtm = lubridate::years(1),
  type_1L_chr = c("all", "main", "change", "zero")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{modifiable_chr}{Modifiable (a character vector)}

\item{pattern_1L_chr}{Pattern (a character vector of length one), Default: '{col}_1_year'}

\item{period_dtm}{Period (a date vector), Default: lubridate::years(1)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("all", "main", "change", "zero")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_test_ds() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update test dataset. The function is called for its side effects and does not return a value.
}
