% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_with_imputed}
\alias{update_with_imputed}
\title{Update with imputed}
\usage{
update_with_imputed(
  project_dss_ls,
  age_1L_chr = "Age",
  employment_1L_chr = "employment_status",
  gender_1L_chr = "gender",
  imputation_args_ls = NULL,
  imputations_fn = mice::mice,
  platform_1L_chr = "platform",
  recode_lup_r3 = make_project_recode_lup(),
  uid_1L_chr = "case_number",
  type_1L_chr = c("modelled", "fixed"),
  what_chr = c("contacts", "outcomes", "overview")
)
}
\arguments{
\item{project_dss_ls}{Project datasets (a list)}

\item{age_1L_chr}{Age (a character vector of length one), Default: 'Age'}

\item{employment_1L_chr}{Employment (a character vector of length one), Default: 'employment_status'}

\item{gender_1L_chr}{Gender (a character vector of length one), Default: 'gender'}

\item{imputation_args_ls}{Imputation arguments (a list), Default: NULL}

\item{imputations_fn}{Imputations (a function), Default: mice::mice}

\item{platform_1L_chr}{Platform (a character vector of length one), Default: 'platform'}

\item{recode_lup_r3}{Recode (a ready4 submodule extension of lookup table), Default: make_project_recode_lup()}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'case_number'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("modelled", "fixed")}

\item{what_chr}{What (a character vector), Default: c("contacts", "outcomes", "overview")}
}
\value{
Imputed datasets (a list)
}
\description{
update_with_imputed() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update with imputed. The function returns Imputed datasets (a list).
}
\keyword{internal}
