% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_batch}
\alias{write_batch}
\title{Write batch}
\usage{
write_batch(
  batch_1L_int,
  add_logic_fn,
  arms_chr,
  base_for_rates_int,
  comparator_fn,
  drop_missing_1L_lgl,
  drop_suffix_1L_chr,
  horizon_dtm,
  inputs_ls,
  intervention_fn,
  iterations_ls,
  modifiable_chr,
  prior_batches_1L_int,
  scale_1L_int,
  seed_1L_int,
  sensitivities_ls,
  start_dtm,
  tfmn_ls,
  utilities_chr,
  variable_unit_1L_chr,
  write_to_1L_chr
)
}
\arguments{
\item{batch_1L_int}{Batch (an integer vector of length one)}

\item{add_logic_fn}{Add logic (a function)}

\item{arms_chr}{Arms (a character vector)}

\item{base_for_rates_int}{Base for rates (an integer vector)}

\item{comparator_fn}{Comparator (a function)}

\item{drop_missing_1L_lgl}{Drop missing (a logical vector of length one)}

\item{drop_suffix_1L_chr}{Drop suffix (a character vector of length one)}

\item{horizon_dtm}{Horizon (a date vector)}

\item{inputs_ls}{Inputs (a list)}

\item{intervention_fn}{Intervention (a function)}

\item{iterations_ls}{Iterations (a list)}

\item{modifiable_chr}{Modifiable (a character vector)}

\item{prior_batches_1L_int}{Prior batches (an integer vector of length one)}

\item{scale_1L_int}{Scale (an integer vector of length one)}

\item{seed_1L_int}{Seed (an integer vector of length one)}

\item{sensitivities_ls}{Sensitivities (a list)}

\item{start_dtm}{Start (a date vector)}

\item{tfmn_ls}{Transformation (a list)}

\item{utilities_chr}{Utilities (a character vector)}

\item{variable_unit_1L_chr}{Variable unit (a character vector of length one)}

\item{write_to_1L_chr}{Write to (a character vector of length one)}
}
\value{
No return value, called for side effects.
}
\description{
write_batch() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write batch. The function is called for its side effects and does not return a value.
}
\keyword{internal}
