% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_project_csvs}
\alias{write_project_csvs}
\title{Write project comma separated variables files}
\usage{
write_project_csvs(
  model_data_ls,
  path_to_private_1L_chr,
  processed_dir_1L_chr,
  divider_1L_chr = "\\\\"
)
}
\arguments{
\item{model_data_ls}{Model data (a list)}

\item{path_to_private_1L_chr}{Path to private (a character vector of length one)}

\item{processed_dir_1L_chr}{Processed directory (a character vector of length one)}

\item{divider_1L_chr}{Divider (a character vector of length one), Default: '\'}
}
\value{
No return value, called for side effects.
}
\description{
write_project_csvs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write project comma separated variables files. The function is called for its side effects and does not return a value.
}
