% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_minutes_event}
\alias{add_minutes_event}
\title{Add minutes event}
\usage{
add_minutes_event(X_Ready4useDyad, minutes_mdl, iterations_int = 1:100L)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{minutes_mdl}{Minutes (a model)}

\item{iterations_int}{Iterations (an integer vector), Default: 1:100L}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_minutes_event() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add minutes event. The function is called for its side effects and does not return a value.
}
\keyword{internal}
