% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_non_helpseekers}
\alias{add_non_helpseekers}
\title{Add non helpseekers}
\usage{
add_non_helpseekers(
  X_Ready4useDyad,
  arms_for_non_helpseeking_chr = character(0)
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{arms_for_non_helpseeking_chr}{Arms for non helpseeking (a character vector), Default: character(0)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_non_helpseekers() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add non helpseekers. The function is called for its side effects and does not return a value.
}
\keyword{internal}
