% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_qalys_sensitivities}
\alias{add_qalys_sensitivities}
\title{Add Quality Adjusted Life Years sensitivities}
\usage{
add_qalys_sensitivities(
  X_Ready4useDyad,
  suffixes_chr = c("_YR1", "_YR1_S1", "_YR1_S2"),
  utility_1L_chr = c("AQoL6D", "CHU9D")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{suffixes_chr}{Suffixes (a character vector), Default: c("_YR1", "_YR1_S1", "_YR1_S2")}

\item{utility_1L_chr}{Utility (a character vector of length one), Default: c("AQoL6D", "CHU9D")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_qalys_sensitivities() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add quality adjusted life years sensitivities. The function is called for its side effects and does not return a value.
}
\keyword{internal}
