% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author-MimicRepos}
\alias{author-MimicRepos}
\alias{author,MimicRepos-method}
\title{Author and save files}
\usage{
\S4method{author}{MimicRepos}(
  x,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N"),
  suffix_1L_chr = "",
  what_1L_chr = c("sim_ws_dirs_chr"),
  ...
)
}
\arguments{
\item{x}{An object of class MimicRepos}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{what_1L_chr}{What (a character vector of length one), Default: c("sim_ws_dirs_chr")}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects.
}
\description{
author method applied to MimicRepos
}
