% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_actives_tb}
\alias{make_actives_tb}
\title{Make actives tibble}
\usage{
make_actives_tb(
  model_data_ls,
  as_tsibble_1L_lgl = FALSE,
  date_end_dtm = NULL,
  date_start_dtm = NULL,
  date_var_1L_chr = make_temporal_vars()
)
}
\arguments{
\item{model_data_ls}{Model data (a list)}

\item{as_tsibble_1L_lgl}{As tsibble (a logical vector of length one), Default: FALSE}

\item{date_end_dtm}{Date end (a date vector), Default: NULL}

\item{date_start_dtm}{Date start (a date vector), Default: NULL}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: make_temporal_vars()}
}
\value{
Actives (an output object of multiple potential types)
}
\description{
make_actives_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make actives tibble. The function returns Actives (an output object of multiple potential types).
}
\keyword{internal}
