% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_2_episode_sequence}
\alias{make_project_2_episode_sequence}
\title{Make project 2 episode sequence}
\usage{
make_project_2_episode_sequence(
  event_nm_1L_chr = "EpisodeofCareSequence",
  outcome_var_1L_chr = "K10",
  start_mdl_1L_chr = "EpisodeStart_mdl",
  use_trigger_1L_chr = "Z",
  validate_schedule_1L_chr = "WaitInDays"
)
}
\arguments{
\item{event_nm_1L_chr}{Event name (a character vector of length one), Default: 'EpisodeofCareSequence'}

\item{outcome_var_1L_chr}{Outcome variable (a character vector of length one), Default: 'K10'}

\item{start_mdl_1L_chr}{Start model (a character vector of length one), Default: 'EpisodeStart_mdl'}

\item{use_trigger_1L_chr}{Use trigger (a character vector of length one), Default: 'Z'}

\item{validate_schedule_1L_chr}{Validate schedule (a character vector of length one), Default: 'WaitInDays'}
}
\value{
X (Model event scheduling and event logic data.)
}
\description{
make_project_2_episode_sequence() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project 2 episode sequence. The function is called for its side effects and does not return a value.
}
\keyword{internal}
