% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_project_results_synthesis}
\alias{make_project_results_synthesis}
\title{Make project results synthesis}
\usage{
make_project_results_synthesis(
  inputs_ls,
  results_ls,
  modifiable_chr = c("treatment_status", "Minutes", "k10", "AQoL6D", "CHU9D"),
  type_1L_chr = c("D", "AB", "C")
)
}
\arguments{
\item{inputs_ls}{Inputs (a list)}

\item{results_ls}{Results (a list)}

\item{modifiable_chr}{Modifiable (a character vector), Default: c("treatment_status", "Minutes", "k10", "AQoL6D", "CHU9D")}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("D", "AB", "C")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
make_project_results_synthesis() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make project results synthesis. The function is called for its side effects and does not return a value.
}
