% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_regression_report}
\alias{make_regression_report}
\title{Make regression report}
\usage{
make_regression_report(
  regressions_ls,
  exclude_int = integer(0),
  model_1L_int = integer(0),
  part_1L_int = integer(0),
  report_1L_chr = c("all", "main", "check", "compare", "confusion", "estimates", "test"),
  rank_1L_lgl = TRUE,
  residual_1L_chr = "normal",
  type_1L_chr = c("candidates", "tests", "models"),
  what_1L_chr = c("AQoL6D", "CHU9D", "K10", "Minutes", "Treatments", "Tx_Waitlist",
    "Tx_Treatment", "Tx_Discharged"),
  var_1L_chr = character(0),
  X_Ready4useDyad = ready4use::Ready4useDyad()
)
}
\arguments{
\item{regressions_ls}{Regressions (a list)}

\item{exclude_int}{Exclude (an integer vector), Default: integer(0)}

\item{model_1L_int}{Model (an integer vector of length one), Default: integer(0)}

\item{part_1L_int}{Part (an integer vector of length one), Default: integer(0)}

\item{report_1L_chr}{Report (a character vector of length one), Default: c("all", "main", "check", "compare", "confusion", "estimates",
"test")}

\item{rank_1L_lgl}{Rank (a logical vector of length one), Default: TRUE}

\item{residual_1L_chr}{Residual (a character vector of length one), Default: 'normal'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("candidates", "tests", "models")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("AQoL6D", "CHU9D", "K10", "Minutes", "Treatments", "Tx_Waitlist",
"Tx_Treatment", "Tx_Discharged")}

\item{var_1L_chr}{Variable (a character vector of length one), Default: character(0)}

\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}
}
\value{
Report (an output object of multiple potential types)
}
\description{
make_regression_report() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make regression report. The function returns Report (an output object of multiple potential types).
}
\keyword{internal}
