% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_regressions_ls}
\alias{make_regressions_ls}
\title{Make regressions list}
\usage{
make_regressions_ls()
}
\value{
Regressions (a list)
}
\description{
make_regressions_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make regressions list. The function returns Regressions (a list).
}
\keyword{internal}
