% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sensitivities_ls}
\alias{make_sensitivities_ls}
\title{Make sensitivities list}
\usage{
make_sensitivities_ls()
}
\value{
Sensitivities (a list)
}
\description{
make_sensitivities_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sensitivities list. The function returns Sensitivities (a list).
}
\keyword{internal}
