% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-MimicArguments}
\alias{manufacture-MimicArguments}
\alias{manufacture,MimicArguments-method}
\alias{manufacture-MimicConfiguration}
\alias{manufacture,MimicConfiguration-method}
\alias{manufacture-MimicEligible}
\alias{manufacture,MimicEligible-method}
\alias{manufacture-MimicRepos}
\alias{manufacture,MimicRepos-method}
\alias{manufacture-MimicPopulation}
\alias{manufacture,MimicPopulation-method}
\alias{manufacture-MimicInputs}
\alias{manufacture,MimicInputs-method}
\alias{manufacture-MimicDerivations}
\alias{manufacture,MimicDerivations-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{MimicArguments}(
  x,
  batch_1L_int = integer(0),
  env_ls = list(),
  what_1L_chr = c("args_ls"),
  X_MimicConfiguration = MimicConfiguration(),
  ...
)

\S4method{manufacture}{MimicConfiguration}(
  x,
  arm_1L_chr = NA_character_,
  batch_1L_int = integer(0),
  draws_tb = NULL,
  extras_ls = list(),
  tx_prefix_1L_chr = character(0),
  type_1L_chr = c("current", "entry"),
  what_1L_chr = c("draws_tb", "args_all", "iterations", "population_ls")
)

\S4method{manufacture}{MimicEligible}(
  x,
  append_ls = list(),
  type_1L_chr = c("filter", "reset"),
  what_1L_chr = "args_ls",
  ...
)

\S4method{manufacture}{MimicRepos}(
  x,
  prefix_1L_chr = character(0),
  return_1L_chr = c("default", "batches", "files"),
  suffix_1L_chr = "",
  type_1L_chr = c("all", "batch_to", "draw_to"),
  what_1L_chr = c("sim_ws_dirs_chr"),
  ...
)

\S4method{manufacture}{MimicPopulation}(x, what_1L_chr = c("population_ls"), ...)

\S4method{manufacture}{MimicInputs}(x, what_1L_chr = c("inputs_ls"))

\S4method{manufacture}{MimicDerivations}(
  x,
  env_ls = list(),
  flatten_1L_lgl = FALSE,
  name_1L_chr = character(0),
  what_1L_chr = c("args_ls"),
  X_MimicConfiguration = MimicConfiguration(),
  ...
)
}
\arguments{
\item{x}{An object of class MimicDerivations}

\item{batch_1L_int}{Batch (an integer vector of length one), Default: integer(0)}

\item{env_ls}{Environment list (a list of environments), Default: list()}

\item{what_1L_chr}{What (a character vector of length one), Default: c("args_ls")}

\item{X_MimicConfiguration}{PARAM_DESCRIPTION, Default: MimicConfiguration()}

\item{...}{Additional arguments}

\item{arm_1L_chr}{Arm (a character vector of length one), Default: 'NA'}

\item{draws_tb}{Draws (a tibble), Default: NULL}

\item{extras_ls}{Extras (a list), Default: list()}

\item{tx_prefix_1L_chr}{Treatment prefix (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("all", "batch_to", "draw_to")}

\item{append_ls}{Append (a list), Default: list()}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: character(0)}

\item{return_1L_chr}{Return (a character vector of length one), Default: c("default", "batches", "files")}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{flatten_1L_lgl}{Flatten (a logical vector of length one), Default: FALSE}

\item{name_1L_chr}{Name (a character vector of length one), Default: character(0)}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
manufacture method applied to MimicArguments

manufacture method applied to MimicConfiguration

manufacture method applied to MimicEligible

manufacture method applied to MimicRepos

manufacture method applied to MimicPopulation

manufacture method applied to MimicInputs

manufacture method applied to MimicDerivations
}
