% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-MimicRepos}
\alias{manufacture-MimicRepos}
\alias{manufacture,MimicRepos-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{MimicRepos}(
  x,
  suffix_1L_chr = "",
  type_1L_chr = c("all", "batch_to"),
  what_1L_chr = c("sim_ws_dirs_chr"),
  ...
)
}
\arguments{
\item{x}{An object of class MimicRepos}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("all", "batch_to")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("sim_ws_dirs_chr")}

\item{...}{Additional arguments}
}
\value{
Object (an output object of multiple potential types)
}
\description{
manufacture method applied to MimicRepos
}
