% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_metamorphose.R
\name{metamorphose-MimicConfiguration}
\alias{metamorphose-MimicConfiguration}
\alias{metamorphose,MimicConfiguration-method}
\title{Metamorphose to a new Ready4Module}
\usage{
\S4method{metamorphose}{MimicConfiguration}(
  x,
  arm_1L_chr = NA_character_,
  batch_1L_int = integer(0),
  draws_tb = NULL,
  tx_prefix_1L_chr = character(0),
  Y_Ready4Module = Ready4Module(),
  ...
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{arm_1L_chr}{Arm (a character vector of length one), Default: 'NA'}

\item{batch_1L_int}{Batch (an integer vector of length one), Default: integer(0)}

\item{draws_tb}{Draws (a tibble), Default: NULL}

\item{tx_prefix_1L_chr}{Treatment prefix (a character vector of length one), Default: character(0)}

\item{...}{Additional arguments}

\item{X_Ready4Module}{PARAM_DESCRIPTION, Default: Ready4Module()}
}
\value{
Object (an output object of multiple potential types)
}
\description{
metamorphose method applied to MimicConfiguration
}
