% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicConfiguration.R
\docType{class}
\name{MimicConfiguration-class}
\alias{MimicConfiguration-class}
\alias{MimicConfiguration}
\title{MimicConfiguration}
\description{
Configuration details for a simulation run.
}
\section{Slots}{

\describe{
\item{\code{arms_chr}}{Arms (a character vector)}

\item{\code{drop_missing_1L_lgl}}{Drop missing (a logical vector of length one)}

\item{\code{drop_suffix_1L_chr}}{Drop suffix (a character vector of length one)}

\item{\code{functions_ls}}{Functions (a list)}

\item{\code{horizon_dtm}}{Horizon (a date vector)}

\item{\code{inputs_ls}}{Inputs (a list)}

\item{\code{iterations_ls}}{Iterations (a list)}

\item{\code{modifiable_chr}}{Modifiable (a character vector)}

\item{\code{prior_batches_1L_int}}{Prior batches (an integer vector of length one)}

\item{\code{seed_1L_int}}{Seed (an integer vector of length one)}

\item{\code{sensitivities_ls}}{Sensitivities (a list)}

\item{\code{start_dtm}}{Start (a date vector)}

\item{\code{utilities_chr}}{Utilities (a character vector)}

\item{\code{x_MimicInputs}}{(an instance of the MimicInputs class)}

\item{\code{x_MimicRepos}}{(an instance of the MimicRepos class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

