% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicEligible.R
\docType{class}
\name{MimicEligible-class}
\alias{MimicEligible-class}
\alias{MimicEligible}
\title{MimicEligible}
\description{
Model event eligibility logic data.
}
\section{Slots}{

\describe{
\item{\code{condition_1L_chr}}{(a character vector of length one)}

\item{\code{functions_ls}}{(a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

