% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_MimicEventsList.R
\docType{class}
\name{MimicEventsList-class}
\alias{MimicEventsList-class}
\alias{MimicEventsList}
\title{MimicEventsList}
\description{
Data on all events included in model.
}
\section{Slots}{

\describe{
\item{\code{events_ls}}{Events (a list)}

\item{\code{last_1L_chr}}{Last (a character vector of length one)}

\item{\code{main_chr}}{Main (a character vector)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

