% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_test_scatter}
\alias{plot_test_scatter}
\title{Plot test scatter}
\usage{
plot_test_scatter(
  X_Ready4useDyad,
  collapse_1L_lgl = F,
  grouping_1L_chr = character(0),
  var_1L_chr,
  old_1L_chr = "Observed",
  type_1L_chr = c("Simulated", "Predicted")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{collapse_1L_lgl}{Collapse (a logical vector of length one), Default: F}

\item{grouping_1L_chr}{Grouping (a character vector of length one), Default: character(0)}

\item{var_1L_chr}{Variable (a character vector of length one)}

\item{old_1L_chr}{Old (a character vector of length one), Default: 'Observed'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("Simulated", "Predicted")}
}
\value{
Scatter (a plot)
}
\description{
plot_test_scatter() is a Plot function that plots data. Specifically, this function implements an algorithm to plot test scatter. The function returns Scatter (a plot).
}
\keyword{internal}
