% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure-MimicConfiguration}
\alias{procure-MimicConfiguration}
\alias{procure,MimicConfiguration-method}
\title{Procure data from a model module}
\usage{
\S4method{procure}{MimicConfiguration}(
  x,
  arm_1L_chr = NA_character_,
  target_1L_chr = character(0),
  what_1L_chr = c("arm")
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{arm_1L_chr}{Arm (a character vector of length one), Default: 'NA'}

\item{target_1L_chr}{Target (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("arm")}
}
\value{
Object (an output object of multiple potential types)
}
\description{
procure method applied to MimicConfiguration
}
