% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_prognosticate.R
\name{prognosticate-MimicConfiguration}
\alias{prognosticate-MimicConfiguration}
\alias{prognosticate,MimicConfiguration-method}
\title{Prognosticate (make predictions) by solving a forward problem}
\usage{
\S4method{prognosticate}{MimicConfiguration}(
  x,
  Y_MimicRepos = MimicRepos(),
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  draws_1L_chr = c("make", "make_batch", "read", "read_batch"),
  options_chr = c("Y", "N"),
  purge_1L_lgl = FALSE,
  suffix_1L_chr = "",
  type_1L_chr = c("NULL", "D", "AB", "C"),
  what_1L_chr = c("all", "batch"),
  unlink_1L_lgl = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{Y_MimicRepos}{PARAM_DESCRIPTION, Default: MimicRepos()}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{draws_1L_chr}{Draws (a character vector of length one), Default: c("make", "make_batch", "read", "read_batch")}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{purge_1L_lgl}{Purge (a logical vector of length one), Default: FALSE}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("NULL", "D", "AB", "C")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("all", "batch")}

\item{unlink_1L_lgl}{Unlink (a logical vector of length one), Default: TRUE}

\item{...}{Additional arguments}
}
\value{
Errors (a list)
}
\description{
prognosticate method applied to MimicConfiguration
}
