% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_prognosticate.R
\name{prognosticate-MimicConfiguration}
\alias{prognosticate-MimicConfiguration}
\alias{prognosticate,MimicConfiguration-method}
\title{Prognosticate (make predictions) by solving a forward problem}
\usage{
\S4method{prognosticate}{MimicConfiguration}(
  x,
  consent_1L_chr = "",
  purge_1L_lgl = FALSE,
  type_1L_chr = c("NULL", "D", "AB", "C"),
  unlink_1L_lgl = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{purge_1L_lgl}{Purge (a logical vector of length one), Default: FALSE}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("NULL", "D", "AB", "C")}

\item{unlink_1L_lgl}{Unlink (a logical vector of length one), Default: TRUE}

\item{...}{Additional arguments}
}
\value{
Errors (a list)
}
\description{
prognosticate method applied to MimicConfiguration
}
