% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew-MimicActive}
\alias{renew-MimicActive}
\alias{renew,MimicActive-method}
\alias{renew-MimicPopulation}
\alias{renew,MimicPopulation-method}
\alias{renew-MimicConfiguration}
\alias{renew,MimicConfiguration-method}
\title{Renew (update) values}
\usage{
\S4method{renew}{MimicActive}(
  x,
  batch_1L_int = integer(0),
  env_ls = list(),
  event_1L_chr = character(0),
  type_1L_chr = c("trigger", "customise", "filter", "schedule"),
  X_MimicConfiguration = MimicConfiguration(),
  ...
)

\S4method{renew}{MimicPopulation}(
  x,
  batch_1L_int = integer(0),
  env_ls = list(),
  event_1L_chr = character(0),
  population_ls = NULL,
  type_1L_chr = c("trigger", "customise", "filter", "event", "reset", "schedule",
    "switch", "transform"),
  use_1L_chr = "Y",
  what_1L_chr = character(0),
  X_MimicConfiguration = MimicConfiguration(),
  ...
)

\S4method{renew}{MimicConfiguration}(x, env_ls = list(), what_1L_chr = c("legacy"), ...)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{batch_1L_int}{Batch (an integer vector of length one), Default: integer(0)}

\item{env_ls}{Environment list (a list of environments), Default: list()}

\item{event_1L_chr}{Event (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("trigger", "customise", "filter", "event", "reset", "schedule",
"switch", "transform")}

\item{X_MimicConfiguration}{PARAM_DESCRIPTION, Default: MimicConfiguration()}

\item{...}{Additional arguments}

\item{population_ls}{Population (a list), Default: NULL}

\item{use_1L_chr}{Use (a character vector of length one), Default: 'Y'}

\item{what_1L_chr}{What (a character vector of length one), Default: c("legacy")}
}
\value{
x (An object of class MimicActive)

x (An object of class MimicPopulation)

x (An object of class MimicConfiguration)
}
\description{
renew method applied to MimicActive

renew method applied to MimicPopulation

renew method applied to MimicConfiguration
}
