% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew-MimicPopulation}
\alias{renew-MimicPopulation}
\alias{renew,MimicPopulation-method}
\alias{renew-MimicConfiguration}
\alias{renew,MimicConfiguration-method}
\title{Renew (update) values}
\usage{
\S4method{renew}{MimicPopulation}(
  x,
  invalid_fn = function(x) (is.na(x) | is.nan(x) | is.null(x) | x == -Inf | x == Inf | x
    < 0),
  schedule_args_ls = list(),
  schedule_fn = NULL,
  population_ls = NULL,
  step_dtm = lubridate::days(0),
  type_1L_chr = c("default", "customise", "schedule", "transform"),
  use_1L_chr = c("Y", "Z"),
  validate_chr = character(0),
  what_1L_chr = character(0),
  X_MimicConfiguration = MimicConfiguration(),
  ...
)

\S4method{renew}{MimicConfiguration}(
  x,
  arm_1L_chr = character(0),
  batch_1L_int = integer(0),
  draws_tb = NULL,
  tx_prefix_1L_chr = character(0),
  type_1L_chr = c("default", "form"),
  what_1L_chr = c("population"),
  ...
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{invalid_fn}{Invalid (a function), Default: function(x) (is.na(x) | is.nan(x) | is.null(x) | x == -Inf |
x == Inf | x < 0)}

\item{schedule_args_ls}{Schedule arguments (a list), Default: list()}

\item{schedule_fn}{Schedule (a function), Default: NULL}

\item{population_ls}{Population (a list), Default: NULL}

\item{step_dtm}{Step (a date vector), Default: lubridate::days(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("default", "form")}

\item{use_1L_chr}{Use (a character vector of length one), Default: c("Y", "Z")}

\item{validate_chr}{Validate (a character vector), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("population")}

\item{X_MimicConfiguration}{PARAM_DESCRIPTION, Default: MimicConfiguration()}

\item{...}{Additional arguments}

\item{arm_1L_chr}{Arm (a character vector of length one), Default: character(0)}

\item{batch_1L_int}{Batch (an integer vector of length one), Default: integer(0)}

\item{draws_tb}{Draws (a tibble), Default: NULL}

\item{tx_prefix_1L_chr}{Treatment prefix (a character vector of length one), Default: character(0)}
}
\value{
x (An object of class MimicPopulation)

x (An object of class MimicConfiguration)
}
\description{
renew method applied to MimicPopulation

renew method applied to MimicConfiguration
}
