% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew-MimicPopulation}
\alias{renew-MimicPopulation}
\alias{renew,MimicPopulation-method}
\alias{renew-MimicConfiguration}
\alias{renew,MimicConfiguration-method}
\title{Renew (update) values}
\usage{
\S4method{renew}{MimicPopulation}(
  x,
  population_ls = NULL,
  type_1L_chr = c("default", "customise", "transform"),
  X_MimicConfiguration = MimicConfiguration(),
  ...
)

\S4method{renew}{MimicConfiguration}(
  x,
  arm_1L_chr = character(0),
  draws_tb = NULL,
  iterations_int = integer(0),
  tx_prefix_1L_chr = character(0),
  type_1L_chr = c("default", "form"),
  what_1L_chr = c("population"),
  ...
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{population_ls}{Population (a list), Default: NULL}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("default", "form")}

\item{X_MimicConfiguration}{PARAM_DESCRIPTION, Default: MimicConfiguration()}

\item{...}{Additional arguments}

\item{arm_1L_chr}{Arm (a character vector of length one), Default: character(0)}

\item{draws_tb}{Draws (a tibble), Default: NULL}

\item{iterations_int}{Iterations (an integer vector), Default: integer(0)}

\item{tx_prefix_1L_chr}{Treatment prefix (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("population")}
}
\value{
x (An object of class MimicPopulation)

x (An object of class MimicConfiguration)
}
\description{
renew method applied to MimicPopulation

renew method applied to MimicConfiguration
}
