% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew-MimicActive}
\alias{renew-MimicActive}
\alias{renew,MimicActive-method}
\alias{renew-MimicPopulation}
\alias{renew,MimicPopulation-method}
\alias{renew-MimicConfiguration}
\alias{renew,MimicConfiguration-method}
\title{Renew (update) values}
\usage{
\S4method{renew}{MimicActive}(
  x,
  batch_1L_int = integer(0),
  env_ls = list(),
  type_1L_chr = c("trigger", "customise", "schedule"),
  X_MimicConfiguration = MimicConfiguration(),
  X_MimicEvent = MimicEvent(),
  ...
)

\S4method{renew}{MimicPopulation}(
  x,
  batch_1L_int = integer(0),
  env_ls = list(),
  population_ls = NULL,
  type_1L_chr = c("trigger", "customise", "schedule", "transform"),
  what_1L_chr = character(0),
  X_MimicConfiguration = MimicConfiguration(),
  X_MimicEvent = MimicEvent(),
  ...
)

\S4method{renew}{MimicConfiguration}(
  x,
  arm_1L_chr = character(0),
  batch_1L_int = integer(0),
  draws_tb = NULL,
  tx_prefix_1L_chr = character(0),
  type_1L_chr = c("default", "form"),
  what_1L_chr = c("population"),
  ...
)
}
\arguments{
\item{x}{An object of class MimicConfiguration}

\item{batch_1L_int}{Batch (an integer vector of length one), Default: integer(0)}

\item{env_ls}{Environment list (a list of environments), Default: list()}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("default", "form")}

\item{X_MimicConfiguration}{PARAM_DESCRIPTION, Default: MimicConfiguration()}

\item{X_MimicEvent}{PARAM_DESCRIPTION, Default: MimicEvent()}

\item{...}{Additional arguments}

\item{population_ls}{Population (a list), Default: NULL}

\item{what_1L_chr}{What (a character vector of length one), Default: c("population")}

\item{arm_1L_chr}{Arm (a character vector of length one), Default: character(0)}

\item{draws_tb}{Draws (a tibble), Default: NULL}

\item{tx_prefix_1L_chr}{Treatment prefix (a character vector of length one), Default: character(0)}
}
\value{
x (An object of class MimicActive)

x (An object of class MimicPopulation)

x (An object of class MimicConfiguration)
}
\description{
renew method applied to MimicActive

renew method applied to MimicPopulation

renew method applied to MimicConfiguration
}
