% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_population_ls}
\alias{update_population_ls}
\title{Update population list}
\usage{
update_population_ls(
  population_ls = NULL,
  X_Ready4useDyad = ready4use::Ready4useDyad(),
  type_1L_chr = c("split", "join", "form"),
  use_1L_chr = c("Y", "Z")
)
}
\arguments{
\item{population_ls}{Population (a list), Default: NULL}

\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("split", "join", "form")}

\item{use_1L_chr}{Use (a character vector of length one), Default: c("Y", "Z")}
}
\value{
Population (a list)
}
\description{
update_population_ls() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update population list. The function returns Population (a list).
}
\keyword{internal}
