% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_batch}
\alias{write_batch}
\title{Write batch}
\usage{
write_batch(
  batch_1L_int,
  arms_chr = character(0),
  comparator_fn = NULL,
  draws_tb = NULL,
  drop_missing_1L_lgl = FALSE,
  drop_suffix_1L_chr = FALSE,
  extra_draws_fn = NULL,
  horizon_dtm = lubridate::years(1),
  inputs_ls = NULL,
  intervention_fn = NULL,
  iterations_ls = NULL,
  modifiable_chr = character(0),
  prior_batches_1L_int = integer(0),
  seed_1L_int = 2001L,
  sensitivities_ls = NULL,
  start_dtm = Sys.Date(),
  tfmn_ls = NULL,
  utilities_chr = character(0),
  write_to_1L_chr,
  X_MimicConfiguration = MimicConfiguration(),
  ...
)
}
\arguments{
\item{batch_1L_int}{Batch (an integer vector of length one)}

\item{arms_chr}{Arms (a character vector), Default: character(0)}

\item{comparator_fn}{Comparator (a function), Default: NULL}

\item{draws_tb}{Draws (a tibble), Default: NULL}

\item{drop_missing_1L_lgl}{Drop missing (a logical vector of length one), Default: FALSE}

\item{drop_suffix_1L_chr}{Drop suffix (a character vector of length one), Default: FALSE}

\item{extra_draws_fn}{Extra draws (a function), Default: NULL}

\item{horizon_dtm}{Horizon (a date vector), Default: lubridate::years(1)}

\item{inputs_ls}{Inputs (a list), Default: NULL}

\item{intervention_fn}{Intervention (a function), Default: NULL}

\item{iterations_ls}{Iterations (a list), Default: NULL}

\item{modifiable_chr}{Modifiable (a character vector), Default: character(0)}

\item{prior_batches_1L_int}{Prior batches (an integer vector of length one), Default: integer(0)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 2001}

\item{sensitivities_ls}{Sensitivities (a list), Default: NULL}

\item{start_dtm}{Start (a date vector), Default: Sys.Date()}

\item{tfmn_ls}{Transformation (a list), Default: NULL}

\item{utilities_chr}{Utilities (a character vector), Default: character(0)}

\item{write_to_1L_chr}{Write to (a character vector of length one)}

\item{X_MimicConfiguration}{PARAM_DESCRIPTION, Default: MimicConfiguration()}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects.
}
\description{
write_batch() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write batch. The function is called for its side effects and does not return a value.
}
\keyword{internal}
